R6 class for package downloads

Download packages with their dependencies, from various sources.

new_pkg_download_proposal(refs, ...)

Arguments

refs
Package names or references. See 'Package references' for the syntax.

...
Additional arguments, passed to pkg_download_proposal$new().

Details

new_pkg_download_proposal() creates a new object from the pkg_download_proposal class, that can be used to look up and download R packages and their dependencies. The advantage of new_pkg_download_proposal() compared to using the pkg_download_proposal constructor directly is that it avoids making pkgdepends a build time dependency.

Typical workflow to download a set of packages:

Create a pkg_download_proposal object with new_pkg_download_proposal().

Resolve all possible dependencies with pkg_download_proposal$resolve().

Download all files with pkg_download_proposal$download().

Get the data about the packages and downloads with pkg_download_proposal$get_downloads().

Methods

Public methods

pkg_download_proposal$new()

pkg_download_proposal$get_refs()

pkg_download_proposal$get_config()

pkg_download_proposal$resolve()

pkg_download_proposal$async_resolve()

pkg_download_proposal$get_resolution()

pkg_download_proposal$download()

pkg_download_proposal$async_download()

pkg_download_proposal$get_downloads()

pkg_download_proposal$stop_for_download_error()

pkg_download_proposal$format()

pkg_download_proposal$print()

pkg_download_proposal$clone()

Method new()

Create a new pkg_download_proposal object. Consider using new_pkg_download_proposal() instead of calling the constructor directly.

The returned object can be used to look up (recursive) dependencies of R packages from various sources, and then to download the package files.

Usage

pkg_download_proposal$new(refs, config = list(), remote_types = NULL)

Arguments

refs

Package names or references. See 'Package references' for the syntax.

config

Configuration options, a named list. See 'Configuration'.

remote_types

Custom remote ref types, this is for advanced use, and experimental currently.

Examples

pdl <- pkg_download_proposal$new("r-lib/pkgdepends")
pdl

Method get_refs()

The package refs that were used to create the pkg_download_proposal object.

Usage

pkg_download_proposal$get_refs()

Returns

A character vector of package refs that were used to create the pkg_download_proposal object.

Examples

pdl <- new_pkg_download_proposal(c("pak", "jsonlite"))
pdl$get_refs()

Method get_config()

Configuration options for the pkg_download_proposal object. See 'Configuration' for details.

Usage

pkg_download_proposal$get_config()

Returns

Named list. See 'Configuration' for the configuration options.

Examples

pdl <- new_pkg_download_proposal("pak")
pdl$get_config()

Method resolve()

Resolve the dependencies of the specified package references. This usually means downloading metadata from CRAN and Bioconductor, unless already cached, and also from GitHub if GitHub refs were included, either directly or indirectly. See 'Dependency resolution' for details.

Usage

pkg_download_proposal$resolve()

Returns

The pkg_download_proposal object itself, invisibly.

Examples

pdl <- new_pkg_download_proposal("pak")
pdl$resolve()
pdl$get_resolution()

Method async_resolve()

The same as resolve(), but asynchronous. This method is for advanced use.

Usage

pkg_download_proposal$async_resolve()

Returns

A deferred value.

Method get_resolution()

Query the result of the dependency resolution. This method can be called after resolve() has completed.

Usage

pkg_download_proposal$get_resolution()

Returns

A pkg_resolution_result object, which is also a tibble. See 'Dependency resolution' for its columns.

Examples

pdl <- new_pkg_download_proposal("r-lib/pkgdepends")
pdl$resolve()
pdl$get_resolution()

Method download()

Download all resolved packages. It uses the package cache in the pkgcache package by default, to avoid downloads if possible.

Usage

pkg_download_proposal$download()

Returns

The pkg_download_proposal object, invisibly.

Examples

pdl <- new_pkg_download_proposal("r-lib/pkgdepends")
pdl$resolve()
pdl$download()
pdl$get_downloads()

Method async_download()

The same as download(), but asynchronous. This method is for advanced use.

Usage

pkg_download_proposal$async_download()

Returns

A deferred value.

Method get_downloads()

Returns the summary of the package downloads.

Usage

pkg_download_proposal$get_downloads()

Returns

A pkg_download_result object, which is a list. See pkg_download_result for details.

Examples

pdl <- new_pkg_download_proposal("pkgload")
pdl$resolve()
pdl$download()
pdl$get_downloads()

Method stop_for_download_error()

Throw and error if the some of the downloads have failed for the most recent pkg_download_proposal$download() call.

Usage

pkg_download_proposal$stop_for_download_error()

Method format()

Format a pkg_download_proposal object, typically for printing.

Usage

pkg_download_proposal$format(...)

Arguments

...

not used currently.

Returns

Nothing. A character vector, each element should be a line in the printout.

Method print()

Prints a pkg_download_proposal object to the screen. The printout includes:

The package refs.

Whether the object has the resolved dependencies.

Whether the resolution had errors.

Whether the downloads were completed.

Whether the downloads had errors.

Advice on which methods to call next.

See the example below.

Usage

pkg_download_proposal$print(...)

Arguments

...

not used currently.

Returns

The pkg_download_proposal object itself, invisibly.

Examples

pdl <- new_pkg_download_proposal("r-lib/pkgdepends")
pdl

pdl$resolve()
pdl

pdl$download()
pdl

Method clone()

The objects of this class are cloneable with this method.

Usage

pkg_download_proposal$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

pdl <- pkg_download_proposal$new("r-lib/pkgdepends")
pdl

#> <pkg_download_proposal>
#> + refs:
#>   - r-lib/pkgdepends
#> (use `$resolve()` to resolve dependencies)

pdl <- new_pkg_download_proposal(c("pak", "jsonlite"))
pdl$get_refs()

#> [1] "pak"      "jsonlite"

pdl <- new_pkg_download_proposal("pak")
pdl$get_config()

#> + config:
#>   - library: 
#>   - cache_dir: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp2cTsaq/file94f761f6629
#>   - package_cache_dir: <default>
#>   - metadata_cache_dir: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp2cTsaq/file94f76c4a55
#>   - platforms: macos, source
#>   - cran-mirror: https://cloud.r-project.org
#>   - dependencies: direct: Depends, Imports, LinkingTo, indirect: Depends, Imports, LinkingTo
#>   - r-versions: 4.0.3
#>   - build-vignettes: FALSE

pdl <- new_pkg_download_proposal("pak")
pdl$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving standard (CRAN/BioC) packages

#>                                                                           

pdl$get_resolution()

#> # A tibble: 86 x 27
#>    ref   type  direct directpkg status package version license needscompilation
#>    <chr> <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 R6    stan… FALSE  FALSE     OK     R6      2.5.0   MIT + … FALSE           
#>  2 asse… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  3 back… stan… FALSE  FALSE     OK     backpo… 1.1.10  GPL-2 … FALSE           
#>  4 base… stan… FALSE  FALSE     OK     base64… 0.1-3   GPL-2 … FALSE           
#>  5 callr stan… FALSE  FALSE     OK     callr   3.5.1   MIT + … FALSE           
#>  6 cli   stan… FALSE  FALSE     OK     cli     2.1.0   MIT + … FALSE           
#>  7 clia… stan… FALSE  FALSE     OK     cliapp  0.1.1   MIT + … FALSE           
#>  8 cray… stan… FALSE  FALSE     OK     crayon  1.3.4   MIT + … FALSE           
#>  9 curl  stan… FALSE  FALSE     OK     curl    4.3     MIT + … FALSE           
#> 10 desc  stan… FALSE  FALSE     OK     desc    1.2.0   MIT + … FALSE           
#> # … with 76 more rows, and 18 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repodir <chr>, target <chr>, deps <list>, mirror <chr>, sources <list>,
#> #   remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, cache_status <chr>

pdl <- new_pkg_download_proposal("r-lib/pkgdepends")
pdl$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving r-lib/pkgdepends

#>  Found  21  deps for  1/1  pkgs [⠙] Resolving standard (CRAN/BioC) packages

#>                                                                            

pdl$get_resolution()

#> # A tibble: 69 x 27
#>    ref   type  direct directpkg status package version license needscompilation
#>    <chr> <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 r-li… gith… TRUE   TRUE      OK     pkgdep… 0.0.0.… MIT + … TRUE            
#>  2 R6    stan… FALSE  FALSE     OK     R6      2.5.0   MIT + … FALSE           
#>  3 asse… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  4 back… stan… FALSE  FALSE     OK     backpo… 1.1.10  GPL-2 … FALSE           
#>  5 callr stan… FALSE  FALSE     OK     callr   3.5.1   MIT + … FALSE           
#>  6 cli   stan… FALSE  FALSE     OK     cli     2.1.0   MIT + … FALSE           
#>  7 cray… stan… FALSE  FALSE     OK     crayon  1.3.4   MIT + … FALSE           
#>  8 curl  stan… FALSE  FALSE     OK     curl    4.3     MIT + … FALSE           
#>  9 desc  stan… FALSE  FALSE     OK     desc    1.2.0   MIT + … FALSE           
#> 10 dige… stan… FALSE  FALSE     OK     digest  0.6.27  GPL (>… FALSE           
#> # … with 59 more rows, and 18 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repodir <chr>, target <glue>, deps <list>, mirror <chr>, sources <list>,
#> #   remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, cache_status <chr>

pdl <- new_pkg_download_proposal("r-lib/pkgdepends")
pdl$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving r-lib/pkgdepends

#>  Found  21  deps for  1/1  pkgs [⠙] Resolving standard (CRAN/BioC) packages

#>                                                                            

pdl$download()

#> ℹ Getting 52 pkgs (11.47 MB) and 1 pkg with unknown size, 16 (7.03 MB) cached

#> 

#>           ⸨             ⸩   0% |  0/53 pkg | Connecting...

#>           ⸨             ⸩   0% |  0/53 pkg | Connecting...

#>           ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends

#>           ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends

#>           ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>           ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of assertthat 0.2.1 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of crayon 1.3.4 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of curl 4.3 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of desc 1.2.0 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of ellipsis 0.3.1 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of fansi 0.4.1 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of filelock 1.0.2 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of lifecycle 0.2.0 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of lpSolve 5.6.15 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of magrittr 1.5 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of pkgconfig 2.0.3 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of prettyunits 1.1.1 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of rappdirs 0.3.1 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of rematch2 2.1.2 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of rprojroot 1.3-2 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of utf8 1.1.4 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>                                                                       

#> ✔ Cached copy of uuid 0.1-4 (macos) is the latest build

#>  150 kB/s ⸨             ⸩   0% |  0/53 pkg | Getting pkgdepends, zip, 

#>  117 kB/s ⸨             ⸩   4% | 17/53 pkg | ETA ~48.9s | Getting pkgdepends, 

#>  348 kB/s ⸨█            ⸩  10% | 17/53 pkg | ETA ~17.4s | Getting pkgdepends, 

#>  485 kB/s ⸨█            ⸩  15% | 17/53 pkg | ETA ~11.8s | Getting pkgdepends, 

#>  508 kB/s ⸨██           ⸩  19% | 17/53 pkg | ETA ~9.9s  | Getting pkgdepends, 

#>  506 kB/s ⸨██           ⸩  17% | 17/53 pkg | ETA ~11.7s | Getting pkgdepends, 

#>  526 kB/s ⸨████         ⸩  36% | 17/53 pkg | ETA ~4.5s  | Getting pkgdepends, 

#>  545 kB/s ⸨████         ⸩  38% | 17/53 pkg | ETA ~4.5s  | Getting pkgdepends, 

#>  599 kB/s ⸨█████        ⸩  42% | 17/53 pkg | ETA ~3.9s  | Getting pkgdepends, 

#>  598 kB/s ⸨█████        ⸩  42% | 17/53 pkg | ETA ~4.1s  | Getting pkgdepends, 

#>  596 kB/s ⸨█████        ⸩  44% | 17/53 pkg | ETA ~4s    | Getting pkgdepends, 

#>  574 kB/s ⸨████         ⸩  37% | 17/53 pkg | ETA ~5.6s  | Getting pkgdepends, 

#>  593 kB/s ⸨████         ⸩  36% | 17/53 pkg | ETA ~6s    | Getting pkgdepends, 

#>                                                                               

#> ✔ Got assertthat 0.2.1 (source) (12.74 kB)

#>  593 kB/s ⸨████         ⸩  36% | 17/53 pkg | ETA ~6s    | Getting pkgdepends, 

#>                                                                               

#> ✔ Got zip 2.1.1 (macos) (211.63 kB)

#>  593 kB/s ⸨████         ⸩  36% | 17/53 pkg | ETA ~6s    | Getting pkgdepends, 

#>                                                                               

#> ✔ Got backports 1.1.10 (source) (19.46 kB)

#>  593 kB/s ⸨████         ⸩  36% | 17/53 pkg | ETA ~6s    | Getting pkgdepends, 

#>  602 kB/s ⸨████         ⸩  37% | 20/53 pkg | ETA ~6.3s  | Getting pkgdepends, 

#>                                                                               

#> ✔ Got cli 2.1.0 (source) (111.41 kB)

#>  602 kB/s ⸨████         ⸩  37% | 20/53 pkg | ETA ~6.3s  | Getting pkgdepends, 

#>                                                                               

#> ✔ Got callr 3.5.1 (source) (77.91 kB)

#>  602 kB/s ⸨████         ⸩  37% | 20/53 pkg | ETA ~6.3s  | Getting pkgdepends, 

#>  677 kB/s ⸨████         ⸩  35% | 22/53 pkg | ETA ~6.9s  | Getting pkgdepends, 

#>                                                                               

#> ✔ Got fansi 0.4.1 (source) (270.91 kB)

#>  677 kB/s ⸨████         ⸩  35% | 22/53 pkg | ETA ~6.9s  | Getting pkgdepends, 

#>                                                                               

#> ✔ Got ellipsis 0.3.1 (source) (7.58 kB)

#>  677 kB/s ⸨████         ⸩  35% | 22/53 pkg | ETA ~6.9s  | Getting pkgdepends, 

#>  692 kB/s ⸨████         ⸩  36% | 24/53 pkg | ETA ~6.8s  | Getting pkgdepends, 

#>  706 kB/s ⸨████         ⸩  37% | 24/53 pkg | ETA ~6.7s  | Getting pkgdepends, 

#>                                                                               

#> ✔ Got crayon 1.3.4 (source) (658.69 kB)

#>  706 kB/s ⸨████         ⸩  37% | 24/53 pkg | ETA ~6.7s  | Getting pkgdepends, 

#>                                                                               

#> ✔ Got filelock 1.0.2 (source) (12.38 kB)

#>  706 kB/s ⸨████         ⸩  37% | 24/53 pkg | ETA ~6.7s  | Getting pkgdepends, 

#>  1.1 MB/s ⸨█████        ⸩  46% | 26/53 pkg | ETA ~3.8s  | Getting pkgdepends, 

#>  1.2 MB/s ⸨██████       ⸩  52% | 26/53 pkg | ETA ~3.1s  | Getting pkgdepends, 

#>                                                                               

#> ✔ Got pkgbuild 1.1.0 (source) (29.85 kB)

#>  1.2 MB/s ⸨██████       ⸩  52% | 26/53 pkg | ETA ~3.1s  | Getting pkgdepends, 

#>                                                                               

#> ✔ Got prettyunits 1.1.1 (source) (10.37 kB)

#>  1.2 MB/s ⸨██████       ⸩  52% | 26/53 pkg | ETA ~3.1s  | Getting pkgdepends, 

#>  1.2 MB/s ⸨███████      ⸩  56% | 28/53 pkg | ETA ~2.7s  | Getting pkgdepends, 

#>                                                                               

#> ✔ Got rappdirs 0.3.1 (source) (12.84 kB)

#>  1.2 MB/s ⸨███████      ⸩  56% | 28/53 pkg | ETA ~2.7s  | Getting pkgdepends, 

#>                                                                               

#> ✔ Got R6 2.5.0 (source) (63.36 kB)

#>  1.2 MB/s ⸨███████      ⸩  56% | 28/53 pkg | ETA ~2.7s  | Getting pkgdepends, 

#>  1.4 MB/s ⸨████████     ⸩  66% | 30/53 pkg | ETA ~1.9s  | Getting pkgdepends, 

#>                                                                               

#> ✔ Got rprojroot 1.3-2 (source) (50.41 kB)

#>  1.4 MB/s ⸨████████     ⸩  66% | 30/53 pkg | ETA ~1.9s  | Getting pkgdepends, 

#>                                                                               

#> ✔ Got pkgconfig 2.0.3 (source) (6.08 kB)

#>  1.4 MB/s ⸨████████     ⸩  66% | 30/53 pkg | ETA ~1.9s  | Getting pkgdepends, 

#>  1.6 MB/s ⸨██████████   ⸩  81% | 32/53 pkg | ETA <1s    | Getting pkgdepends, 

#>                                                                               

#> ✔ Got uuid 0.1-4 (source) (49.54 kB)

#>  1.6 MB/s ⸨██████████   ⸩  81% | 32/53 pkg | ETA <1s    | Getting pkgdepends, 

#>  1.8 MB/s ⸨████████████ ⸩  93% | 33/53 pkg | ETA <1s    | Getting pkgdepends, 

#>                                                                               

#> ✔ Got rematch2 2.1.2 (source) (13.37 kB)

#>  1.8 MB/s ⸨████████████ ⸩  93% | 33/53 pkg | ETA <1s    | Getting pkgdepends, 

#>                                                                               

#> ✔ Got glue 1.4.2 (source) (99.05 kB)

#>  1.8 MB/s ⸨████████████ ⸩  93% | 33/53 pkg | ETA <1s    | Getting pkgdepends, 

#>  2.3 MB/s ⸨█████████████⸩ 101% | 35/53 pkg | ETA <1s    | Getting pkgdepends, 

#>                                                                               

#> ✔ Got desc 1.2.0 (source) (65.61 kB)

#>  2.3 MB/s ⸨█████████████⸩ 101% | 35/53 pkg | ETA <1s    | Getting pkgdepends, 

#>  2.2 MB/s ⸨█████████████⸩ 101% | 36/53 pkg | ETA <1s    | Getting pkgdepends  

#>                                                                             

#> ✔ Got pillar 1.4.6 (source) (113.88 kB)

#>  2.2 MB/s ⸨█████████████⸩ 101% | 36/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got ps 1.4.0 (source) (110.71 kB)

#>  2.2 MB/s ⸨█████████████⸩ 101% | 36/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got zip 2.1.1 (source) (112.86 kB)

#>  2.2 MB/s ⸨█████████████⸩ 101% | 36/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got withr 2.3.0 (source) (91.44 kB)

#>  2.2 MB/s ⸨█████████████⸩ 101% | 36/53 pkg | ETA <1s    | Getting pkgdepends

#>  2.1 MB/s ⸨█████████████⸩ 101% | 40/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got lifecycle 0.2.0 (source) (164.46 kB)

#>  2.1 MB/s ⸨█████████████⸩ 101% | 40/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got digest 0.6.27 (source) (164.37 kB)

#>  2.1 MB/s ⸨█████████████⸩ 101% | 40/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got pkgcache 1.1.1 (source) (131.82 kB)

#>  2.1 MB/s ⸨█████████████⸩ 101% | 40/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got processx 3.4.4 (source) (134.70 kB)

#>  2.1 MB/s ⸨█████████████⸩ 101% | 40/53 pkg | ETA <1s    | Getting pkgdepends

#>  2.0 MB/s ⸨█████████████⸩ 101% | 44/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got magrittr 1.5 (source) (200.50 kB)

#>  2.0 MB/s ⸨█████████████⸩ 101% | 44/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got utf8 1.1.4 (source) (218.88 kB)

#>  2.0 MB/s ⸨█████████████⸩ 101% | 44/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got tibble 3.0.4 (source) (255.46 kB)

#>  2.0 MB/s ⸨█████████████⸩ 101% | 44/53 pkg | ETA <1s    | Getting pkgdepends

#>  2.0 MB/s ⸨█████████████⸩ 101% | 47/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got rlang 0.4.8 (source) (847.52 kB)

#>  2.0 MB/s ⸨█████████████⸩ 101% | 47/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got lpSolve 5.6.15 (source) (467.65 kB)

#>  2.0 MB/s ⸨█████████████⸩ 101% | 47/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got vctrs 0.3.4 (source) (995.08 kB)

#>  2.0 MB/s ⸨█████████████⸩ 101% | 47/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got curl 4.3 (source) (673.78 kB)

#>  2.0 MB/s ⸨█████████████⸩ 101% | 47/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got jsonlite 1.7.1 (source) (1.06 MB)

#>  2.0 MB/s ⸨█████████████⸩ 101% | 47/53 pkg | ETA <1s    | Getting pkgdepends

#>  1.9 MB/s ⸨█████████████⸩ 101% | 52/53 pkg | ETA <1s    | Getting pkgdepends

#>  1.9 MB/s ⸨█████████████⸩ 102% | 52/53 pkg | ETA <1s    | Getting pkgdepends

#>  1.8 MB/s ⸨█████████████⸩ 102% | 52/53 pkg | ETA <1s    | Getting pkgdepends

#>  1.9 MB/s ⸨█████████████⸩ 102% | 52/53 pkg | ETA <1s    | Getting pkgdepends

#>  1.9 MB/s ⸨█████████████⸩ 103% | 52/53 pkg | ETA <1s    | Getting pkgdepends

#>  1.8 MB/s ⸨█████████████⸩ 103% | 52/53 pkg | ETA <1s    | Getting pkgdepends

#>  1.8 MB/s ⸨█████████████⸩ 103% | 52/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Got pkgdepends 0.0.0.9007 (source) (269.30 kB)

#>  1.8 MB/s ⸨█████████████⸩ 103% | 52/53 pkg | ETA <1s    | Getting pkgdepends

#>                                                                             

#> ✔ Downloaded 36 packages (7.79 MB) in 6.5s

pdl$get_downloads()

#> # A tibble: 69 x 32
#>    ref   type  direct directpkg status package version license needscompilation
#>    <chr> <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 r-li… gith… TRUE   TRUE      OK     pkgdep… 0.0.0.… MIT + … TRUE            
#>  2 R6    stan… FALSE  FALSE     OK     R6      2.5.0   MIT + … FALSE           
#>  3 asse… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  4 back… stan… FALSE  FALSE     OK     backpo… 1.1.10  GPL-2 … FALSE           
#>  5 callr stan… FALSE  FALSE     OK     callr   3.5.1   MIT + … FALSE           
#>  6 cli   stan… FALSE  FALSE     OK     cli     2.1.0   MIT + … FALSE           
#>  7 cray… stan… FALSE  FALSE     OK     crayon  1.3.4   MIT + … FALSE           
#>  8 curl  stan… FALSE  FALSE     OK     curl    4.3     MIT + … FALSE           
#>  9 desc  stan… FALSE  FALSE     OK     desc    1.2.0   MIT + … FALSE           
#> 10 dige… stan… FALSE  FALSE     OK     digest  0.6.27  GPL (>… FALSE           
#> # … with 59 more rows, and 23 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repodir <chr>, target <glue>, deps <list>, mirror <chr>, sources <list>,
#> #   remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, cache_status <chr>, fulltarget <chr>,
#> #   fulltarget_tree <chr>, download_status <chr>, download_error <list>,
#> #   file_size <dbl>

pdl <- new_pkg_download_proposal("pkgload")
pdl$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving standard (CRAN/BioC) packages

#>                                                                           

pdl$download()

#> ℹ Getting 10 pkgs (1.97 MB), 26 (5.83 MB) cached

#> 

#>           ⸨             ⸩   0% |  0/10 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/10 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/10 pkg | ETA ??s  | Getting pkgload, 

#>                                                                          

#> ✔ Cached copy of assertthat 0.2.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/10 pkg | ETA ??s  | Getting pkgload, 

#>                                                                          

#> ✔ Cached copy of crayon 1.3.4 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/10 pkg | ETA ??s  | Getting pkgload, 

#>                                                                          

#> ✔ Cached copy of desc 1.2.0 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/10 pkg | ETA ??s  | Getting pkgload, 

#>                                                                          

#> ✔ Cached copy of fansi 0.4.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/10 pkg | ETA ??s  | Getting pkgload, 

#>                                                                          

#> ✔ Cached copy of prettyunits 1.1.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/10 pkg | ETA ??s  | Getting pkgload, 

#>                                                                          

#> ✔ Cached copy of rprojroot 1.3-2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/10 pkg | ETA ??s  | Getting pkgload, 

#>  594 kB/s ⸨███████████  ⸩  88% |  6/10 pkg | ETA <1s    | Getting pkgload, 

#>                                                                            

#> ✔ Got pkgload 1.1.0 (macos) (150.88 kB)

#>  594 kB/s ⸨███████████  ⸩  88% |  6/10 pkg | ETA <1s    | Getting pkgload, 

#>                                                                            

#> ✔ Got rstudioapi 0.11 (macos) (257.65 kB)

#>  594 kB/s ⸨███████████  ⸩  88% |  6/10 pkg | ETA <1s    | Getting pkgload, 

#>                                                                            

#> ✔ Got pkgload 1.1.0 (source) (58.05 kB)

#>  594 kB/s ⸨███████████  ⸩  88% |  6/10 pkg | ETA <1s    | Getting pkgload, 

#>                                                                            

#> ✔ Got rstudioapi 0.11 (source) (98.08 kB)

#>  594 kB/s ⸨███████████  ⸩  88% |  6/10 pkg | ETA <1s    | Getting pkgload, 

#>  672 kB/s ⸨██████████████⸩ 111% | 10/10 pkg | ETA <1s    | Getting pkgload and 

#>                                                                                

#> ✔ Downloaded 4 packages (564.65 kB) in 863ms

pdl$get_downloads()

#> # A tibble: 36 x 32
#>    ref   type  direct directpkg status package version license needscompilation
#>    <chr> <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 R6    stan… FALSE  FALSE     OK     R6      2.5.0   MIT + … FALSE           
#>  2 asse… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  3 back… stan… FALSE  FALSE     OK     backpo… 1.1.10  GPL-2 … FALSE           
#>  4 callr stan… FALSE  FALSE     OK     callr   3.5.1   MIT + … FALSE           
#>  5 cli   stan… FALSE  FALSE     OK     cli     2.1.0   MIT + … FALSE           
#>  6 cray… stan… FALSE  FALSE     OK     crayon  1.3.4   MIT + … FALSE           
#>  7 desc  stan… FALSE  FALSE     OK     desc    1.2.0   MIT + … FALSE           
#>  8 fansi stan… FALSE  FALSE     OK     fansi   0.4.1   GPL (>… FALSE           
#>  9 glue  stan… FALSE  FALSE     OK     glue    1.4.2   MIT + … FALSE           
#> 10 pkgb… stan… FALSE  FALSE     OK     pkgbui… 1.1.0   GPL-3   FALSE           
#> # … with 26 more rows, and 23 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repodir <chr>, target <chr>, deps <list>, mirror <chr>, sources <list>,
#> #   remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, cache_status <chr>, fulltarget <chr>,
#> #   fulltarget_tree <chr>, download_status <chr>, download_error <list>,
#> #   file_size <dbl>

pdl <- new_pkg_download_proposal("r-lib/pkgdepends")
pdl

#> <pkg_download_proposal>
#> + refs:
#>   - r-lib/pkgdepends
#> (use `$resolve()` to resolve dependencies)

pdl$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving r-lib/pkgdepends

#>  Found  21  deps for  1/1  pkgs [⠙] Resolving standard (CRAN/BioC) packages

#>                                                                            

pdl

#> <pkg_download_proposal>
#> + refs:
#>   - r-lib/pkgdepends
#> + has resolution (+34 dependencies)
#> (use `$download()` to download packages)
#> (use `$get_resolution()` to see resolution results)

pdl$download()

#> ℹ Getting 17 pkgs (3.95 MB) and 1 pkg with unknown size, 51 (14.56 MB) cached

#> 

#>           ⸨             ⸩   0% |  0/18 pkg | Connecting...

#>           ⸨             ⸩   0% |  0/18 pkg | Connecting...

#>                                                           

#> ✔ Cached copy of pkgdepends 0.0.0.9007 (source) is the latest build

#>           ⸨             ⸩   0% |  0/18 pkg | Connecting...

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of assertthat 0.2.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of crayon 1.3.4 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of curl 4.3 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of desc 1.2.0 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of ellipsis 0.3.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of fansi 0.4.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of filelock 1.0.2 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of lifecycle 0.2.0 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of lpSolve 5.6.15 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of magrittr 1.5 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of pkgconfig 2.0.3 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of prettyunits 1.1.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of rappdirs 0.3.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of rematch2 2.1.2 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of rprojroot 1.3-2 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of utf8 1.1.4 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of uuid 0.1-4 (macos) is the latest build

#>           ⸨             ⸩   0% |  1/18 pkg | ETA ??s  | Connecting...

#>           ⸨█████████████⸩ 100% | 18/18 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ No downloads needed, all packages are cached

pdl

#> <pkg_download_proposal>
#> + refs:
#>   - r-lib/pkgdepends
#> + has resolution (+34 dependencies)
#> + has downloads
#> (use `$get_resolution()` to see resolution results)
#> (use `$get_downloads()` to get download data)