The download progress bar

This is a short summary of the design of the download progress bar.

Details

Package sizes

One difficulty here is that we don't always know the sizes of the files we are downloading. We basically have three cases for downloads:

For some, we know the sizes from the extra CRAN metadata. This is usually correct, unless CRAN has changed the file on the web server. If we get a size from the web server, then we'll use that.

For some, we do not know the size beforehand, but the web server sends us the correct size, together with the first chunk of data. This is typically for CRAN packages that we don't have metadata about, Bioconductor packages, or GitHub packages for which GitHub already knows the size.

For some, we do not know the size, and the web server does not tell us the size, either. This is typically for GH packages that GH is creating on the fly.

So ideally we would have a different progress bar when the total size if known.

Possible information to show:

Visual progress bar

Number of packages: done / in-progress / total

Number of bytes: done / total (if known)

ETA

Download rate, i.e. x kB/s

Elapsed time

Percent of packages done

Percent of bytes done

Spinner to show that we are still working, although updating ETA is good as well...

Events, e.g. got a package, downloading, or starting...

All package sizes are known

230 kB/s (#####      ) 45% | 14/56 pkgs | ETA ~34s | Getting dplyr

Rate is informative, serves as an "alive" indicator as well.

Bar is visual.

Percent is the best quantitative measure.

Done/total packages is informative, not so much for the download but the whole installation.

ETA is informative.

Event is nice, and it fills the space...

We don't show current/total bytes, shoulwd we? Quite wide...

Some package sizes are unknown

230 kB/s (#####      ) 45% | 14/56 pkgs | Getting dplyr

Percent is for the packages, not bytes.

We can't really show ETA.

Events

This is quite simple currently, we just print the packages we are "getting" currently, i.e. the packages we received data for in the last time slot.

Rate (download "speed")

To calculate the rate we need to record when the various data chunks have arrived. We only do this with a resolution of 1s. We create an environment called chunks, and for every second since the downloads have started we record the number of bytes that arrived in that second.

When calculating the rate, we take the data for the last 4 seconds, and normalize it to per second, while taking into account that the data for the last (current) second is incomplete.

In the first 3 seconds we use all the available data.