Installation plans

An installation plan contains all data that is needed to install a set of package files. It is usually created from an installation proposal with solving the dependencies and downloading the package files.

Details

It is also possible to create an installation plan a different way. An installation plan object must be a data frame, with at least the following columns:

package: The name of the package.

type: The type of the package reference.

binary: Whether the package is a binary package.

file: Full path to the package file or directory.

dependencies: A list column that lists the names of the dependent packages for each package.

needscompilation: Whether the package needs compilation. This should be FALSE for binary packages.

For installation plans created via pkg_installation_proposal, the plan contains all columns from pkg_download_result objects, and some additional ones:

library: the library the package is supposed to be installed to.

direct: whether the package was directly requested or it is installed as a dependency.

vignettes: whether the vignettes need to be (re)built.

packaged: whether R CMD build was already called for the package.

See also

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp9XlBCJ/file8f872347

pdi$resolve()

#> 

#> ✔ Updated metadata database: 4.39 MB in 12 files.

#> 

#> ℹ Updating metadata database

#> ✔ Updating metadata database ... done

#> 

pdi$solve()
pdi$download()

#> ℹ Getting 40 pkgs (27.79 MB) and 2 pkgs with unknown sizes

#> 

#> ✔ Got R6 2.5.0 (macos) (82.45 kB)

#> 

#> ✔ Got assertthat 0.2.1 (macos) (52.57 kB)

#> 

#> ✔ Got base64enc 0.1-3 (macos) (31.70 kB)

#> 

#> ✔ Got callr 3.5.1 (macos) (385.44 kB)

#> 

#> ✔ Got cli 2.3.1 (macos) (469.80 kB)

#> 

#> ✔ Got cliapp 0.1.1 (macos) (235.74 kB)

#> 

#> ✔ Got desc 1.3.0 (macos) (527.78 kB)

#> 

#> ✔ Got digest 0.6.27 (macos) (300.37 kB)

#> 

#> ✔ Got ellipsis 0.3.1 (macos) (33.50 kB)

#> 

#> ✔ Got fansi 0.4.2 (macos) (212.15 kB)

#> 

#> ✔ Got crayon 1.4.1 (macos) (139.92 kB)

#> 

#> ✔ Got glue 1.4.2 (macos) (139.02 kB)

#> 

#> ✔ Got lifecycle 1.0.0 (macos) (93.31 kB)

#> 

#> ✔ Got filelock 1.0.2 (macos) (26.73 kB)

#> 

#> ✔ Got hms 1.0.0 (macos) (97.75 kB)

#> 

#> ✔ Got magrittr 2.0.1 (macos) (224.85 kB)

#> 

#> ✔ Got pkgconfig 2.0.3 (macos) (17.74 kB)

#> 

#> ✔ Got prettyunits 1.1.1 (macos) (34.70 kB)

#> 

#> ✔ Got rappdirs 0.3.3 (macos) (45.62 kB)

#> 

#> ✔ Got rematch2 2.1.2 (macos) (44.81 kB)

#> 

#> ✔ Got pkgbuild 1.2.0 (macos) (140.59 kB)

#> 

#> ✔ Got uuid 0.1-4 (macos) (27.81 kB)

#> 

#> ✔ Got progress 1.2.2 (macos) (83.46 kB)

#> 

#> ✔ Got rprojroot 2.0.2 (macos) (96.29 kB)

#> 

#> ✔ Got utf8 1.1.4 (macos) (195.53 kB)

#> 

#> ✔ Got prettycode 1.1.0 (macos) (242.92 kB)

#> 

#> ✔ Got stringr 1.4.0 (macos) (210.65 kB)

#> 

#> ✔ Got withr 2.4.1 (macos) (205.90 kB)

#> 

#> ✔ Got processx 3.4.5 (macos) (279.43 kB)

#> 

#> ✔ Got ps 1.6.0 (macos) (285.36 kB)

#> 

#> ✔ Got pak 0.1.2.1 (macos) (404.31 kB)

#> 

#> ✔ Got jsonlite 1.7.2 (macos) (492.07 kB)

#> 

#> ✔ Got pillar 1.5.1 (macos) (951.07 kB)

#> 

#> ✔ Got selectr 0.4-2 (macos) (485.31 kB)

#> 

#> ✔ Got pkgcache 1.2.0 (macos) (654.50 kB)

#> 

#> ✔ Got curl 4.3 (macos) (741.52 kB)

#> 

#> ✔ Got rlang 0.4.10 (macos) (1.33 MB)

#> 

#> ✔ Got tibble 3.1.0 (macos) (803.56 kB)

#> 

#> ✔ Got lpSolve 5.6.15 (macos) (1.07 MB)

#> 

#> ✔ Got vctrs 0.3.6 (macos) (1.40 MB)

#> 

#> ✔ Got stringi 1.5.3 (macos) (13.64 MB)

#> 

#> ✔ Got xml2 1.3.2 (macos) (2.34 MB)

#> 

#> ✔ Downloaded 42 packages (29.27 MB) in 15.4s

pdi$get_install_plan()

#> # A tibble: 42 x 44
#>    ref    type  direct directpkg status package version license needscompilation
#>    <chr>  <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 R6     stan… FALSE  FALSE     OK     R6      2.5.0   MIT + … FALSE           
#>  2 asser… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  3 base6… stan… FALSE  FALSE     OK     base64… 0.1-3   GPL-2 … FALSE           
#>  4 callr  stan… FALSE  FALSE     OK     callr   3.5.1   MIT + … FALSE           
#>  5 cli    stan… FALSE  FALSE     OK     cli     2.3.1   MIT + … FALSE           
#>  6 cliapp stan… FALSE  FALSE     OK     cliapp  0.1.1   MIT + … FALSE           
#>  7 crayon stan… FALSE  FALSE     OK     crayon  1.4.1   MIT + … FALSE           
#>  8 curl   stan… FALSE  FALSE     OK     curl    4.3     MIT + … FALSE           
#>  9 desc   stan… FALSE  FALSE     OK     desc    1.3.0   MIT + … FALSE           
#> 10 digest stan… FALSE  FALSE     OK     digest  0.6.27  GPL (>… FALSE           
#> # … with 32 more rows, and 35 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repotype <chr>, repodir <chr>, target <chr>, deps <list>, mirror <chr>,
#> #   sources <list>, remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, params <list>, cache_status <chr>, lib_status <chr>,
#> #   old_version <chr>, new_version <chr>, fulltarget <chr>,
#> #   fulltarget_tree <chr>, download_status <chr>, download_error <list>,
#> #   file_size <dbl>, library <chr>, binary <lgl>, dependencies <I<list>>,
#> #   installed <chr>, vignettes <lgl>, packaged <lgl>, file <chr>