Installation plans

An installation plan contains all data that is needed to install a set of package files. It is usually created from an installation proposal with solving the dependencies and downloading the package files.

Details

It is also possible to create an installation plan a different way. An installation plan object must be a data frame, with at least the following columns:

package: The name of the package.

type: The type of the package reference.

binary: Whether the package is a binary package.

file: Full path to the package file or directory.

dependencies: A list column that lists the names of the dependent packages for each package.

needscompilation: Whether the package needs compilation. This should be FALSE for binary packages.

For installation plans created via pkg_installation_proposal, the plan contains all columns from pkg_download_result objects, and some additional ones:

library: the library the package is supposed to be installed to.

direct: whether the package was directly requested or it is installed as a dependency.

vignettes: whether the vignettes need to be (re)built.

packaged: whether R CMD build was already called for the package.

See also

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpTn2Adk/file6e129390de7

pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠹ Updating metadata [0/12] | Downloading [3.75 MB / 4.05 MB]

#> ⠸ Updating metadata [0/12] | Downloading [3.84 MB / 4.05 MB]

#> ⠼ Updating metadata [0/12] | Downloading [3.87 MB / 4.05 MB]

#> ⠴ Updating metadata [0/12] | Downloading [3.87 MB / 4.05 MB]

#> ⠦ Updating metadata [0/12] | Downloading [3.90 MB / 4.05 MB]

#> ⠧ Updating metadata [0/12] | Downloading [3.94 MB / 4.05 MB]

#> ⠇ Updating metadata [0/12] | Downloading [3.97 MB / 4.05 MB]

#> ⠏ Updating metadata [0/12] | Downloading [3.97 MB / 4.05 MB]

#> ⠋ Updating metadata [0/12] | Downloading [3.98 MB / 4.03 MB]

#> ⠙ Updating metadata [0/12] | Downloading [4.03 MB / 4.03 MB]

#> ⠹ Updating metadata [0/12] | Downloading [4.03 MB / 4.03 MB]

#> ✔ Updated metadata: 4.03 MB in 12 files.                    

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving standard (CRAN/BioC) packages

#>                                                                           

pdi$solve()
pdi$download()

#> ℹ Getting 43 pkgs (28.56 MB)

#> 

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Getting R6, assertthat,

#>  226 kB/s ⸨             ⸩   6% |  0/43 pkg | ETA ~1m    | Getting R6,          

#>  283 kB/s ⸨             ⸩   7% |  0/43 pkg | ETA ~1m    | Getting cliapp, 

#>  118 kB/s ⸨█            ⸩   8% |  0/43 pkg | ETA ~3m    | Getting cliapp, curl,

#>  125 kB/s ⸨█            ⸩   9% |  0/43 pkg | ETA ~3m    | Getting cliapp, curl,

#>  131 kB/s ⸨█            ⸩   9% |  0/43 pkg | ETA ~3m    | Getting cliapp, curl,

#>  121 kB/s ⸨█            ⸩  10% |  0/43 pkg | ETA ~3m    | Getting cliapp, curl,

#>   93 kB/s ⸨█            ⸩  10% |  0/43 pkg | ETA ~4m    | Getting cliapp, curl,

#>   93 kB/s ⸨█            ⸩  11% |  0/43 pkg | ETA ~4m    | Getting cliapp, curl,

#>  116 kB/s ⸨█            ⸩  12% |  0/43 pkg | ETA ~3m    | Getting cliapp, curl,

#>   97 kB/s ⸨█            ⸩  12% |  0/43 pkg | ETA ~4m    | Getting cliapp, curl,

#>  114 kB/s ⸨█            ⸩  13% |  0/43 pkg | ETA ~3m    | Getting cliapp, curl,

#>  169 kB/s ⸨█            ⸩  14% |  0/43 pkg | ETA ~2m    | Getting cliapp, curl,

#>  198 kB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~2m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got R6 2.4.1 (macos) (57.55 kB)

#>  198 kB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~2m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got assertthat 0.2.1 (macos) (52.57 kB)

#>  198 kB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~2m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got backports 1.1.9 (macos) (75.40 kB)

#>  198 kB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~2m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got base64enc 0.1-3 (macos) (31.70 kB)

#>  198 kB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~2m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got callr 3.4.3 (macos) (385.70 kB)

#>  198 kB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~2m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got cli 2.0.2 (macos) (395.50 kB)

#>  198 kB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~2m    | Getting cliapp, curl,

#>  288 kB/s ⸨██           ⸩  17% |  6/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got crayon 1.3.4 (macos) (748.90 kB)

#>  288 kB/s ⸨██           ⸩  17% |  6/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got ellipsis 0.3.1 (macos) (33.50 kB)

#>  288 kB/s ⸨██           ⸩  17% |  6/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got filelock 1.0.2 (macos) (26.73 kB)

#>  288 kB/s ⸨██           ⸩  17% |  6/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>  282 kB/s ⸨██           ⸩  18% |  9/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got prettyunits 1.1.1 (macos) (34.70 kB)

#>  282 kB/s ⸨██           ⸩  18% |  9/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>  141 kB/s ⸨██           ⸩  18% | 10/43 pkg | ETA ~2m    | Getting cliapp, curl,

#>   98 kB/s ⸨██           ⸩  19% | 10/43 pkg | ETA ~3m    | Getting cliapp, curl,

#>  120 kB/s ⸨██           ⸩  20% | 10/43 pkg | ETA ~3m    | Getting cliapp, curl,

#>  143 kB/s ⸨██           ⸩  21% | 10/43 pkg | ETA ~2m    | Getting cliapp, curl,

#>  204 kB/s ⸨██           ⸩  22% | 10/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>  235 kB/s ⸨██           ⸩  22% | 10/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>  252 kB/s ⸨███          ⸩  23% | 10/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got uuid 0.1-4 (macos) (27.81 kB)

#>  252 kB/s ⸨███          ⸩  23% | 10/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got rematch2 2.1.2 (macos) (44.81 kB)

#>  252 kB/s ⸨███          ⸩  23% | 10/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>                                                                                

#> ✔ Got pkgconfig 2.0.3 (macos) (17.74 kB)

#>  252 kB/s ⸨███          ⸩  23% | 10/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>  338 kB/s ⸨███          ⸩  24% | 13/43 pkg | ETA ~1m    | Getting cliapp, curl,

#>  410 kB/s ⸨███          ⸩  25% | 13/43 pkg | ETA ~52.1s | Getting cliapp, curl,

#>                                                                                

#> ✔ Got lifecycle 0.2.0 (macos) (91.60 kB)

#>  410 kB/s ⸨███          ⸩  25% | 13/43 pkg | ETA ~52.1s | Getting cliapp, curl,

#>                                                                                

#> ✔ Got hms 0.5.3 (macos) (103.86 kB)

#>  410 kB/s ⸨███          ⸩  25% | 13/43 pkg | ETA ~52.1s | Getting cliapp, curl,

#>  482 kB/s ⸨███          ⸩  26% | 15/43 pkg | ETA ~43.7s | Getting cliapp, curl,

#>                                                                                

#> ✔ Got rprojroot 1.3-2 (macos) (79.15 kB)

#>  482 kB/s ⸨███          ⸩  26% | 15/43 pkg | ETA ~43.7s | Getting cliapp, curl,

#>  545 kB/s ⸨███          ⸩  27% | 16/43 pkg | ETA ~38.1s | Getting curl,        

#>                                                                         

#> ✔ Got progress 1.2.2 (macos) (83.46 kB)

#>  545 kB/s ⸨███          ⸩  27% | 16/43 pkg | ETA ~38.1s | Getting curl, 

#>  633 kB/s ⸨███          ⸩  29% | 17/43 pkg | ETA ~32.1s | Getting curl, 

#>  691 kB/s ⸨███          ⸩  30% | 17/43 pkg | ETA ~29s   | Getting curl, 

#>                                                                         

#> ✔ Got glue 1.4.1 (macos) (138.29 kB)

#>  691 kB/s ⸨███          ⸩  30% | 17/43 pkg | ETA ~29s   | Getting curl, 

#>  1.0 MB/s ⸨████         ⸩  32% | 18/43 pkg | ETA ~18.8s | Getting curl, 

#>                                                                         

#> ✔ Got pillar 1.4.6 (macos) (176.43 kB)

#>  1.0 MB/s ⸨████         ⸩  32% | 18/43 pkg | ETA ~18.8s | Getting curl, 

#>  1.2 MB/s ⸨████         ⸩  35% | 19/43 pkg | ETA ~15.6s | Getting curl, 

#>                                                                         

#> ✔ Got magrittr 1.5 (macos) (153.23 kB)

#>  1.2 MB/s ⸨████         ⸩  35% | 19/43 pkg | ETA ~15.6s | Getting curl, 

#>  1.2 MB/s ⸨████         ⸩  35% | 20/43 pkg | ETA ~15.3s | Getting curl, 

#>                                                                         

#> ✔ Got rappdirs 0.3.1 (macos) (145.34 kB)

#>  1.2 MB/s ⸨████         ⸩  35% | 20/43 pkg | ETA ~15.3s | Getting curl, 

#>  1.5 MB/s ⸨█████        ⸩  39% | 21/43 pkg | ETA ~11.8s | Getting curl, 

#>                                                                         

#> ✔ Got pkgbuild 1.1.0 (macos) (139.36 kB)

#>  1.5 MB/s ⸨█████        ⸩  39% | 21/43 pkg | ETA ~11.8s | Getting curl, 

#>                                                                         

#> ✔ Got desc 1.2.0 (macos) (285.23 kB)

#>  1.5 MB/s ⸨█████        ⸩  39% | 21/43 pkg | ETA ~11.8s | Getting curl, 

#>                                                                         

#> ✔ Got digest 0.6.25 (macos) (246.06 kB)

#>  1.5 MB/s ⸨█████        ⸩  39% | 21/43 pkg | ETA ~11.8s | Getting curl, 

#>  1.9 MB/s ⸨█████        ⸩  45% | 24/43 pkg | ETA ~8.3s  | Getting curl, 

#>                                                                         

#> ✔ Got cliapp 0.1.0 (macos) (234.28 kB)

#>  1.9 MB/s ⸨█████        ⸩  45% | 24/43 pkg | ETA ~8.3s  | Getting curl, 

#>                                                                         

#> ✔ Got fansi 0.4.1 (macos) (210.70 kB)

#>  1.9 MB/s ⸨█████        ⸩  45% | 24/43 pkg | ETA ~8.3s  | Getting curl, 

#>  2.1 MB/s ⸨██████       ⸩  49% | 26/43 pkg | ETA ~7s    | Getting jsonlite, 

#>                                                                             

#> ✔ Got withr 2.2.0 (macos) (205.74 kB)

#>  2.1 MB/s ⸨██████       ⸩  49% | 26/43 pkg | ETA ~7s    | Getting jsonlite, 

#>                                                                             

#> ✔ Got ps 1.3.4 (macos) (217.61 kB)

#>  2.1 MB/s ⸨██████       ⸩  49% | 26/43 pkg | ETA ~7s    | Getting jsonlite, 

#>                                                                             

#> ✔ Got prettycode 1.1.0 (macos) (242.92 kB)

#>  2.1 MB/s ⸨██████       ⸩  49% | 26/43 pkg | ETA ~7s    | Getting jsonlite, 

#>  2.4 MB/s ⸨███████      ⸩  54% | 29/43 pkg | ETA ~5.5s  | Getting jsonlite, 

#>                                                                             

#> ✔ Got stringr 1.4.0 (macos) (210.65 kB)

#>  2.4 MB/s ⸨███████      ⸩  54% | 29/43 pkg | ETA ~5.5s  | Getting jsonlite, 

#>                                                                             

#> ✔ Got utf8 1.1.4 (macos) (195.53 kB)

#>  2.4 MB/s ⸨███████      ⸩  54% | 29/43 pkg | ETA ~5.5s  | Getting jsonlite, 

#>  3.2 MB/s ⸨███████      ⸩  57% | 31/43 pkg | ETA ~3.8s  | Getting stringi and 

#>                                                                               

#> ✔ Got processx 3.4.3 (macos) (278.45 kB)

#>  3.2 MB/s ⸨███████      ⸩  57% | 31/43 pkg | ETA ~3.8s  | Getting stringi and 

#>  6.1 MB/s ⸨█████████████⸩ 100% | 32/43 pkg | ETA <1s    | Getting stringi     

#>                                                                          

#> ✔ Got pak 0.1.2 (macos) (403.75 kB)

#>  6.1 MB/s ⸨█████████████⸩ 100% | 32/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got tibble 3.0.3 (macos) (394.30 kB)

#>  6.1 MB/s ⸨█████████████⸩ 100% | 32/43 pkg | ETA <1s    | Getting stringi

#>  5.8 MB/s ⸨█████████████⸩ 100% | 34/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got selectr 0.4-2 (macos) (485.31 kB)

#>  5.8 MB/s ⸨█████████████⸩ 100% | 34/43 pkg | ETA <1s    | Getting stringi

#>  7.2 MB/s ⸨█████████████⸩ 100% | 35/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got pkgcache 1.1.0 (macos) (548.74 kB)

#>  7.2 MB/s ⸨█████████████⸩ 100% | 35/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got curl 4.3 (macos) (741.52 kB)

#>  7.2 MB/s ⸨█████████████⸩ 100% | 35/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got vctrs 0.3.2 (macos) (1.27 MB)

#>  7.2 MB/s ⸨█████████████⸩ 100% | 35/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got jsonlite 1.7.0 (macos) (1.12 MB)

#>  7.2 MB/s ⸨█████████████⸩ 100% | 35/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got lpSolve 5.6.15 (macos) (1.07 MB)

#>  7.2 MB/s ⸨█████████████⸩ 100% | 35/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got rlang 0.4.7 (macos) (1.21 MB)

#>  7.2 MB/s ⸨█████████████⸩ 100% | 35/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got xml2 1.3.2 (macos) (2.34 MB)

#>  7.2 MB/s ⸨█████████████⸩ 100% | 35/43 pkg | ETA <1s    | Getting stringi

#>  6.9 MB/s ⸨█████████████⸩ 100% | 42/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got stringi 1.4.6 (macos) (13.62 MB)

#>  6.9 MB/s ⸨█████████████⸩ 100% | 42/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Downloaded 43 packages (28.57 MB) in 46.2s

pdi$get_install_plan()

#> # A tibble: 43 x 42
#>    ref   type  direct directpkg status package version license needscompilation
#>    <chr> <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 R6    stan… FALSE  FALSE     OK     R6      2.4.1   MIT + … FALSE           
#>  2 asse… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  3 back… stan… FALSE  FALSE     OK     backpo… 1.1.9   GPL-2 … FALSE           
#>  4 base… stan… FALSE  FALSE     OK     base64… 0.1-3   GPL-2 … FALSE           
#>  5 callr stan… FALSE  FALSE     OK     callr   3.4.3   MIT + … FALSE           
#>  6 cli   stan… FALSE  FALSE     OK     cli     2.0.2   MIT + … FALSE           
#>  7 clia… stan… FALSE  FALSE     OK     cliapp  0.1.0   MIT + … FALSE           
#>  8 cray… stan… FALSE  FALSE     OK     crayon  1.3.4   MIT + … FALSE           
#>  9 curl  stan… FALSE  FALSE     OK     curl    4.3     MIT + … FALSE           
#> 10 desc  stan… FALSE  FALSE     OK     desc    1.2.0   MIT + … FALSE           
#> # … with 33 more rows, and 33 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repodir <chr>, target <chr>, deps <list>, mirror <chr>, sources <list>,
#> #   remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, cache_status <chr>, lib_status <chr>, old_version <chr>,
#> #   new_version <chr>, fulltarget <chr>, fulltarget_tree <chr>,
#> #   download_status <chr>, download_error <list>, file_size <dbl>,
#> #   library <chr>, binary <lgl>, dependencies <I<list>>, installed <chr>,
#> #   vignettes <lgl>, packaged <lgl>, file <chr>