Installation plans

An installation plan contains all data that is needed to install a set of package files. It is usually created from an installation proposal with solving the dependencies and downloading the package files.

Details

It is also possible to create an installation plan a different way. An installation plan object must be a data frame, with at least the following columns:

package: The name of the package.

type: The type of the package reference.

binary: Whether the package is a binary package.

file: Full path to the package file or directory.

dependencies: A list column that lists the names of the dependent packages for each package.

needscompilation: Whether the package needs compilation. This should be FALSE for binary packages.

For installation plans created via pkg_installation_proposal, the plan contains all columns from pkg_download_result objects, and some additional ones:

library: the library the package is supposed to be installed to.

direct: whether the package was directly requested or it is installed as a dependency.

vignettes: whether the vignettes need to be (re)built.

packaged: whether R CMD build was already called for the package.

See also

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp5MApS8/file64c57b9d007

pdi$resolve()

#> 

#> ✔ Updated metadata: 4.12 MB in 12 files.

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving standard (CRAN/BioC) packages

#>                                                                           

pdi$solve()
pdi$download()

#> ℹ Getting 42 pkgs (28.29 MB) and 1 pkg with unknown size

#> 

#>           ⸨             ⸩   0% |  0/43 pkg | Connecting...

#>           ⸨             ⸩   0% |  0/43 pkg | Connecting...

#>           ⸨             ⸩   0% |  0/43 pkg | Connecting...

#>           ⸨             ⸩   0% |  0/43 pkg | Getting R6, assertthat, backports,

#>  898 kB/s ⸨             ⸩   3% |  0/43 pkg | ETA ~31s   | Getting R6,          

#>  870 kB/s ⸨             ⸩   3% |  0/43 pkg | ETA ~31.9s | Getting cli, cliapp, 

#>  870 kB/s ⸨             ⸩   4% |  0/43 pkg | ETA ~31.8s | Getting cli, crayon, 

#>  833 kB/s ⸨             ⸩   4% |  0/43 pkg | ETA ~33.1s | Getting cli, crayon, 

#>  800 kB/s ⸨             ⸩   4% |  0/43 pkg | ETA ~34.4s | Getting cli, crayon, 

#>  757 kB/s ⸨             ⸩   4% |  0/43 pkg | ETA ~36.4s | Getting cli, crayon, 

#>  667 kB/s ⸨             ⸩   4% |  0/43 pkg | ETA ~41.1s | Getting cli, crayon, 

#>  651 kB/s ⸨             ⸩   4% |  0/43 pkg | ETA ~42.1s | Getting cli, crayon, 

#>  640 kB/s ⸨             ⸩   5% |  0/43 pkg | ETA ~42.7s | Getting cli, crayon, 

#>  659 kB/s ⸨             ⸩   5% |  0/43 pkg | ETA ~41.3s | Getting cli, crayon, 

#>  668 kB/s ⸨             ⸩   6% |  0/43 pkg | ETA ~40.5s | Getting cli, crayon, 

#>  719 kB/s ⸨             ⸩   6% |  0/43 pkg | ETA ~37.3s | Getting cli, crayon, 

#>  714 kB/s ⸨             ⸩   7% |  0/43 pkg | ETA ~37.5s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got assertthat 0.2.1 (macos) (52.57 kB)

#>  714 kB/s ⸨             ⸩   7% |  0/43 pkg | ETA ~37.5s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got base64enc 0.1-3 (macos) (31.70 kB)

#>  714 kB/s ⸨             ⸩   7% |  0/43 pkg | ETA ~37.5s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got R6 2.4.1 (macos) (57.55 kB)

#>  714 kB/s ⸨             ⸩   7% |  0/43 pkg | ETA ~37.5s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got backports 1.1.10 (macos) (78.36 kB)

#>  714 kB/s ⸨             ⸩   7% |  0/43 pkg | ETA ~37.5s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got callr 3.5.1 (macos) (385.44 kB)

#>  714 kB/s ⸨             ⸩   7% |  0/43 pkg | ETA ~37.5s | Getting cli, crayon, 

#>  669 kB/s ⸨             ⸩   7% |  5/43 pkg | ETA ~40s   | Getting cli, crayon, 

#>                                                                                

#> ✔ Got cliapp 0.1.1 (macos) (235.74 kB)

#>  669 kB/s ⸨             ⸩   7% |  5/43 pkg | ETA ~40s   | Getting cli, crayon, 

#>                                                                                

#> ✔ Got filelock 1.0.2 (macos) (26.73 kB)

#>  669 kB/s ⸨             ⸩   7% |  5/43 pkg | ETA ~40s   | Getting cli, crayon, 

#>  643 kB/s ⸨             ⸩   7% |  7/43 pkg | ETA ~41.5s | Getting cli, crayon, 

#>  635 kB/s ⸨             ⸩   7% |  7/43 pkg | ETA ~42s   | Getting cli, crayon, 

#>  617 kB/s ⸨             ⸩   7% |  7/43 pkg | ETA ~43.1s | Getting cli, crayon, 

#>  604 kB/s ⸨             ⸩   7% |  7/43 pkg | ETA ~44s   | Getting cli, crayon, 

#>  599 kB/s ⸨█            ⸩   8% |  7/43 pkg | ETA ~44.2s | Getting cli, crayon, 

#>  588 kB/s ⸨█            ⸩   8% |  7/43 pkg | ETA ~45s   | Getting cli, crayon, 

#>  580 kB/s ⸨█            ⸩   8% |  7/43 pkg | ETA ~45.5s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got hms 0.5.3 (macos) (103.86 kB)

#>  580 kB/s ⸨█            ⸩   8% |  7/43 pkg | ETA ~45.5s | Getting cli, crayon, 

#>  475 kB/s ⸨█            ⸩   8% |  8/43 pkg | ETA ~55.4s | Getting cli, crayon, 

#>  464 kB/s ⸨█            ⸩   8% |  8/43 pkg | ETA ~56.7s | Getting cli, crayon, 

#>  449 kB/s ⸨█            ⸩   8% |  8/43 pkg | ETA ~58.6s | Getting cli, crayon, 

#>  439 kB/s ⸨█            ⸩   9% |  8/43 pkg | ETA ~59.8s | Getting cli, crayon, 

#>  434 kB/s ⸨█            ⸩   9% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  425 kB/s ⸨█            ⸩   9% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  421 kB/s ⸨█            ⸩   9% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  414 kB/s ⸨█            ⸩   9% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  411 kB/s ⸨█            ⸩   9% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  407 kB/s ⸨█            ⸩   9% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  406 kB/s ⸨█            ⸩  10% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  419 kB/s ⸨█            ⸩  10% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  417 kB/s ⸨█            ⸩  10% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  417 kB/s ⸨█            ⸩  10% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  325 kB/s ⸨█            ⸩  10% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  337 kB/s ⸨█            ⸩  11% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>                                                                                

#> ✔ Got pkgconfig 2.0.3 (macos) (17.74 kB)

#>  337 kB/s ⸨█            ⸩  11% |  8/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  329 kB/s ⸨█            ⸩  11% |  9/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>                                                                                

#> ✔ Got rematch2 2.1.2 (macos) (44.81 kB)

#>  329 kB/s ⸨█            ⸩  11% |  9/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  325 kB/s ⸨█            ⸩  11% | 10/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  317 kB/s ⸨█            ⸩  11% | 10/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  313 kB/s ⸨█            ⸩  11% | 10/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  297 kB/s ⸨█            ⸩  11% | 10/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  289 kB/s ⸨█            ⸩  11% | 10/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  288 kB/s ⸨█            ⸩  11% | 10/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  285 kB/s ⸨█            ⸩  12% | 10/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  282 kB/s ⸨█            ⸩  12% | 10/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  276 kB/s ⸨█            ⸩  12% | 10/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>                                                                                

#> ✔ Got uuid 0.1-4 (macos) (27.81 kB)

#>  276 kB/s ⸨█            ⸩  12% | 10/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  275 kB/s ⸨█            ⸩  12% | 11/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>                                                                                

#> ✔ Got prettyunits 1.1.1 (macos) (34.70 kB)

#>  275 kB/s ⸨█            ⸩  12% | 11/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  269 kB/s ⸨█            ⸩  12% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  272 kB/s ⸨█            ⸩  12% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  269 kB/s ⸨█            ⸩  12% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  263 kB/s ⸨█            ⸩  12% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  258 kB/s ⸨█            ⸩  12% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  247 kB/s ⸨█            ⸩  12% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  243 kB/s ⸨█            ⸩  12% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  242 kB/s ⸨█            ⸩  12% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  199 kB/s ⸨█            ⸩  12% | 12/43 pkg | ETA ~2m    | Getting cli, crayon, 

#>  199 kB/s ⸨█            ⸩  12% | 12/43 pkg | ETA ~2m    | Getting cli, crayon, 

#>  201 kB/s ⸨█            ⸩  13% | 12/43 pkg | ETA ~2m    | Getting cli, crayon, 

#>  204 kB/s ⸨█            ⸩  13% | 12/43 pkg | ETA ~2m    | Getting cli, crayon, 

#>  207 kB/s ⸨█            ⸩  13% | 12/43 pkg | ETA ~2m    | Getting cli, crayon, 

#>  231 kB/s ⸨█            ⸩  13% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  231 kB/s ⸨█            ⸩  13% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  234 kB/s ⸨█            ⸩  13% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  284 kB/s ⸨█            ⸩  14% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  302 kB/s ⸨█            ⸩  14% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  369 kB/s ⸨██           ⸩  15% | 12/43 pkg | ETA ~1m    | Getting cli, crayon, 

#>  421 kB/s ⸨██           ⸩  16% | 12/43 pkg | ETA ~57s   | Getting cli, crayon, 

#>  473 kB/s ⸨██           ⸩  17% | 12/43 pkg | ETA ~50.2s | Getting cli, crayon, 

#>  560 kB/s ⸨██           ⸩  19% | 12/43 pkg | ETA ~41.6s | Getting cli, crayon, 

#>  683 kB/s ⸨██           ⸩  19% | 12/43 pkg | ETA ~34s   | Getting cli, crayon, 

#>  692 kB/s ⸨██           ⸩  20% | 12/43 pkg | ETA ~33.3s | Getting cli, crayon, 

#>  701 kB/s ⸨██           ⸩  20% | 12/43 pkg | ETA ~32.7s | Getting cli, crayon, 

#>  709 kB/s ⸨██           ⸩  21% | 12/43 pkg | ETA ~32.1s | Getting cli, crayon, 

#>  812 kB/s ⸨██           ⸩  23% | 12/43 pkg | ETA ~27.4s | Getting cli, crayon, 

#>  809 kB/s ⸨██           ⸩  23% | 12/43 pkg | ETA ~27.3s | Getting cli, crayon, 

#>  1.1 MB/s ⸨███          ⸩  24% | 12/43 pkg | ETA ~20s   | Getting cli, crayon, 

#>  1.1 MB/s ⸨███          ⸩  25% | 12/43 pkg | ETA ~19.6s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got ellipsis 0.3.1 (macos) (33.50 kB)

#>  1.1 MB/s ⸨███          ⸩  25% | 12/43 pkg | ETA ~19.6s | Getting cli, crayon, 

#>  1.1 MB/s ⸨███          ⸩  26% | 13/43 pkg | ETA ~19s   | Getting cli, crayon, 

#>  1.1 MB/s ⸨███          ⸩  27% | 13/43 pkg | ETA ~18.8s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got progress 1.2.2 (macos) (83.46 kB)

#>  1.1 MB/s ⸨███          ⸩  27% | 13/43 pkg | ETA ~18.8s | Getting cli, crayon, 

#>  1.3 MB/s ⸨███          ⸩  28% | 14/43 pkg | ETA ~15.3s | Getting cli, crayon, 

#>  1.3 MB/s ⸨███          ⸩  29% | 14/43 pkg | ETA ~15.1s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got rprojroot 1.3-2 (macos) (79.15 kB)

#>  1.3 MB/s ⸨███          ⸩  29% | 14/43 pkg | ETA ~15.1s | Getting cli, crayon, 

#>  1.3 MB/s ⸨███          ⸩  29% | 15/43 pkg | ETA ~15.3s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got lifecycle 0.2.0 (macos) (91.60 kB)

#>  1.3 MB/s ⸨███          ⸩  29% | 15/43 pkg | ETA ~15.3s | Getting cli, crayon, 

#>  1.3 MB/s ⸨███          ⸩  30% | 16/43 pkg | ETA ~15.3s | Getting cli, crayon, 

#>  1.3 MB/s ⸨███          ⸩  30% | 16/43 pkg | ETA ~15.2s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got pillar 1.4.6 (macos) (176.43 kB)

#>  1.3 MB/s ⸨███          ⸩  30% | 16/43 pkg | ETA ~15.2s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got fansi 0.4.1 (macos) (210.70 kB)

#>  1.3 MB/s ⸨███          ⸩  30% | 16/43 pkg | ETA ~15.2s | Getting cli, crayon, 

#>  1.3 MB/s ⸨████         ⸩  31% | 18/43 pkg | ETA ~15.3s | Getting cli, crayon, 

#>  1.2 MB/s ⸨████         ⸩  32% | 18/43 pkg | ETA ~15.9s | Getting cli, crayon, 

#>  1.2 MB/s ⸨████         ⸩  33% | 18/43 pkg | ETA ~16s   | Getting cli, crayon, 

#>                                                                                

#> ✔ Got magrittr 1.5 (macos) (153.23 kB)

#>  1.2 MB/s ⸨████         ⸩  33% | 18/43 pkg | ETA ~16s   | Getting cli, crayon, 

#>                                                                                

#> ✔ Got pkgbuild 1.1.0 (macos) (139.36 kB)

#>  1.2 MB/s ⸨████         ⸩  33% | 18/43 pkg | ETA ~16s   | Getting cli, crayon, 

#>                                                                                

#> ✔ Got glue 1.4.2 (macos) (139.02 kB)

#>  1.2 MB/s ⸨████         ⸩  33% | 18/43 pkg | ETA ~16s   | Getting cli, crayon, 

#>                                                                                

#> ✔ Got rappdirs 0.3.1 (macos) (145.34 kB)

#>  1.2 MB/s ⸨████         ⸩  33% | 18/43 pkg | ETA ~16s   | Getting cli, crayon, 

#>  1.3 MB/s ⸨████         ⸩  35% | 22/43 pkg | ETA ~14.3s | Getting cli, crayon, 

#>  1.3 MB/s ⸨████         ⸩  36% | 22/43 pkg | ETA ~13.8s | Getting cli, crayon, 

#>  1.3 MB/s ⸨████         ⸩  37% | 22/43 pkg | ETA ~13.6s | Getting cli, crayon, 

#>                                                                                

#> ✔ Got ps 1.4.0 (macos) (251.68 kB)

#>  1.3 MB/s ⸨████         ⸩  37% | 22/43 pkg | ETA ~13.6s | Getting cli, crayon, 

#>  1.3 MB/s ⸨████         ⸩  38% | 23/43 pkg | ETA ~13.8s | Getting crayon, curl,

#>                                                                                

#> ✔ Got prettycode 1.1.0 (macos) (242.92 kB)

#>  1.3 MB/s ⸨████         ⸩  38% | 23/43 pkg | ETA ~13.8s | Getting crayon, curl,

#>                                                                                

#> ✔ Got digest 0.6.25 (macos) (246.06 kB)

#>  1.3 MB/s ⸨████         ⸩  38% | 23/43 pkg | ETA ~13.8s | Getting crayon, curl,

#>                                                                                

#> ✔ Got withr 2.3.0 (macos) (205.49 kB)

#>  1.3 MB/s ⸨████         ⸩  38% | 23/43 pkg | ETA ~13.8s | Getting crayon, curl,

#>                                                                                

#> ✔ Got stringr 1.4.0 (macos) (210.65 kB)

#>  1.3 MB/s ⸨████         ⸩  38% | 23/43 pkg | ETA ~13.8s | Getting crayon, curl,

#>  1.3 MB/s ⸨█████        ⸩  39% | 27/43 pkg | ETA ~13.3s | Getting crayon, curl,

#>                                                                                

#> ✔ Got utf8 1.1.4 (macos) (195.53 kB)

#>  1.3 MB/s ⸨█████        ⸩  39% | 27/43 pkg | ETA ~13.3s | Getting crayon, curl,

#>  1.3 MB/s ⸨█████        ⸩  40% | 28/43 pkg | ETA ~13.1s | Getting crayon,      

#>  1.3 MB/s ⸨█████        ⸩  40% | 28/43 pkg | ETA ~13s   | Getting crayon, 

#>                                                                           

#> ✔ Got processx 3.4.4 (macos) (278.50 kB)

#>  1.3 MB/s ⸨█████        ⸩  40% | 28/43 pkg | ETA ~13s   | Getting crayon, 

#>  1.3 MB/s ⸨█████        ⸩  41% | 29/43 pkg | ETA ~12.9s | Getting crayon, 

#>                                                                           

#> ✔ Got desc 1.2.0 (macos) (285.23 kB)

#>  1.3 MB/s ⸨█████        ⸩  41% | 29/43 pkg | ETA ~12.9s | Getting crayon, 

#>  1.2 MB/s ⸨█████        ⸩  42% | 30/43 pkg | ETA ~13.8s | Getting crayon, 

#>  1.3 MB/s ⸨█████        ⸩  43% | 30/43 pkg | ETA ~13.1s | Getting crayon, 

#>  1.3 MB/s ⸨█████        ⸩  44% | 30/43 pkg | ETA ~12.2s | Getting jsonlite, 

#>  1.3 MB/s ⸨█████        ⸩  46% | 30/43 pkg | ETA ~11.8s | Getting jsonlite, 

#>  1.4 MB/s ⸨██████       ⸩  47% | 30/43 pkg | ETA ~10.7s | Getting jsonlite, 

#>  1.4 MB/s ⸨██████       ⸩  49% | 30/43 pkg | ETA ~10.4s | Getting jsonlite, 

#>                                                                             

#> ✔ Got pak 0.1.2 (macos) (403.75 kB)

#>  1.4 MB/s ⸨██████       ⸩  49% | 30/43 pkg | ETA ~10.4s | Getting jsonlite, 

#>                                                                             

#> ✔ Got cli 2.1.0 (macos) (406.23 kB)

#>  1.4 MB/s ⸨██████       ⸩  49% | 30/43 pkg | ETA ~10.4s | Getting jsonlite, 

#>  1.5 MB/s ⸨██████       ⸩  51% | 32/43 pkg | ETA ~9.5s  | Getting jsonlite, 

#>                                                                             

#> ✔ Got tibble 3.0.4 (macos) (393.07 kB)

#>  1.5 MB/s ⸨██████       ⸩  51% | 32/43 pkg | ETA ~9.5s  | Getting jsonlite, 

#>  1.6 MB/s ⸨███████      ⸩  54% | 33/43 pkg | ETA ~8.1s  | Getting stringi,  

#>  1.7 MB/s ⸨███████      ⸩  57% | 33/43 pkg | ETA ~7.3s  | Getting stringi, 

#>                                                                            

#> ✔ Got curl 4.3 (macos) (741.52 kB)

#>  1.7 MB/s ⸨███████      ⸩  57% | 33/43 pkg | ETA ~7.3s  | Getting stringi, 

#>                                                                            

#> ✔ Got selectr 0.4-2 (macos) (485.31 kB)

#>  1.7 MB/s ⸨███████      ⸩  57% | 33/43 pkg | ETA ~7.3s  | Getting stringi, 

#>  1.8 MB/s ⸨███████      ⸩  61% | 35/43 pkg | ETA ~6.2s  | Getting stringi and 

#>  2.0 MB/s ⸨████████     ⸩  65% | 35/43 pkg | ETA ~5.1s  | Getting stringi and 

#>                                                                               

#> ✔ Got pkgcache 1.1.1 (macos) (551.45 kB)

#>  2.0 MB/s ⸨████████     ⸩  65% | 35/43 pkg | ETA ~5.1s  | Getting stringi and 

#>  2.2 MB/s ⸨████████     ⸩  69% | 36/43 pkg | ETA ~4.1s  | Getting stringi and 

#>  2.4 MB/s ⸨█████████    ⸩  72% | 36/43 pkg | ETA ~3.3s  | Getting stringi and 

#>                                                                               

#> ✔ Got crayon 1.3.4 (macos) (748.90 kB)

#>  2.4 MB/s ⸨█████████    ⸩  72% | 36/43 pkg | ETA ~3.3s  | Getting stringi and 

#>  2.5 MB/s ⸨█████████    ⸩  76% | 37/43 pkg | ETA ~2.7s  | Getting stringi and 

#>                                                                               

#> ✔ Got lpSolve 5.6.15 (macos) (1.07 MB)

#>  2.5 MB/s ⸨█████████    ⸩  76% | 37/43 pkg | ETA ~2.7s  | Getting stringi and 

#>  2.8 MB/s ⸨██████████   ⸩  81% | 38/43 pkg | ETA ~2s    | Getting stringi and 

#>  2.8 MB/s ⸨██████████   ⸩  83% | 38/43 pkg | ETA ~1.7s  | Getting stringi and 

#>  3.6 MB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi     

#>                                                                          

#> ✔ Got rlang 0.4.8 (macos) (1.21 MB)

#>  3.6 MB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got jsonlite 1.7.1 (macos) (1.12 MB)

#>  3.6 MB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got vctrs 0.3.4 (macos) (1.31 MB)

#>  3.6 MB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi

#>  3.6 MB/s ⸨█████████████⸩ 100% | 41/43 pkg | ETA <1s    | Getting stringi

#>  3.6 MB/s ⸨█████████████⸩ 100% | 41/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got xml2 1.3.2 (macos) (2.34 MB)

#>  3.6 MB/s ⸨█████████████⸩ 100% | 41/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got stringi 1.5.3 (macos) (13.64 MB)

#>  3.6 MB/s ⸨█████████████⸩ 100% | 41/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Downloaded 43 packages (28.69 MB) in 22.1s

pdi$get_install_plan()

#> # A tibble: 43 x 42
#>    ref   type  direct directpkg status package version license needscompilation
#>    <chr> <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 R6    stan… FALSE  FALSE     OK     R6      2.4.1   MIT + … FALSE           
#>  2 asse… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  3 back… stan… FALSE  FALSE     OK     backpo… 1.1.10  GPL-2 … FALSE           
#>  4 base… stan… FALSE  FALSE     OK     base64… 0.1-3   GPL-2 … FALSE           
#>  5 callr stan… FALSE  FALSE     OK     callr   3.5.1   MIT + … FALSE           
#>  6 cli   stan… FALSE  FALSE     OK     cli     2.1.0   MIT + … FALSE           
#>  7 clia… stan… FALSE  FALSE     OK     cliapp  0.1.1   MIT + … FALSE           
#>  8 cray… stan… FALSE  FALSE     OK     crayon  1.3.4   MIT + … FALSE           
#>  9 curl  stan… FALSE  FALSE     OK     curl    4.3     MIT + … FALSE           
#> 10 desc  stan… FALSE  FALSE     OK     desc    1.2.0   MIT + … FALSE           
#> # … with 33 more rows, and 33 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repodir <chr>, target <chr>, deps <list>, mirror <chr>, sources <list>,
#> #   remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, cache_status <chr>, lib_status <chr>, old_version <chr>,
#> #   new_version <chr>, fulltarget <chr>, fulltarget_tree <chr>,
#> #   download_status <chr>, download_error <list>, file_size <dbl>,
#> #   library <chr>, binary <lgl>, dependencies <I<list>>, installed <chr>,
#> #   vignettes <lgl>, packaged <lgl>, file <chr>