Installation plans

An installation plan contains all data that is needed to install a set of package files. It is usually created from an installation proposal with solving the dependencies and downloading the package files.

Details

It is also possible to create an installation plan a different way. An installation plan object must be a data frame, with at least the following columns:

package: The name of the package.

type: The type of the package reference.

binary: Whether the package is a binary package.

file: Full path to the package file or directory.

dependencies: A list column that lists the names of the dependent packages for each package.

needscompilation: Whether the package needs compilation. This should be FALSE for binary packages.

For installation plans created via pkg_installation_proposal, the plan contains all columns from pkg_download_result objects, and some additional ones:

library: the library the package is supposed to be installed to.

direct: whether the package was directly requested or it is installed as a dependency.

vignettes: whether the vignettes need to be (re)built.

packaged: whether R CMD build was already called for the package.

See also

Examples

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$download()
pdi$get_install_plan()
}) # examplesIf

#> > pdi <- new_pkg_installation_proposal("pak", config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp3d1tuA/file90c44dd2857

#> > pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [32.74 kB / 162.44 kB]

#> ⠹ Updating metadata [0/12] | Downloading [1.75 MB / 1.91 MB]   

#> ⠸ Updating metadata [0/12] | Downloading [1.89 MB / 1.89 MB]

#> ⠼ Updating metadata [0/12] | Downloading [1.99 MB / 4.00 MB]

#> ⠴ Updating metadata [0/12] | Downloading [2.02 MB / 4.00 MB]

#> ⠦ Updating metadata [0/12] | Downloading [2.10 MB / 4.00 MB]

#> ⠧ Updating metadata [0/12] | Downloading [2.25 MB / 4.00 MB]

#> ⠇ Updating metadata [0/12] | Downloading [2.28 MB / 4.00 MB]

#> ⠏ Updating metadata [0/12] | Downloading [2.47 MB / 4.00 MB]

#> ⠋ Updating metadata [0/12] | Downloading [2.58 MB / 4.00 MB]

#> ⠙ Updating metadata [0/12] | Downloading [2.73 MB / 4.00 MB]

#> ⠹ Updating metadata [0/12] | Downloading [2.84 MB / 4.00 MB]

#> ⠸ Updating metadata [0/12] | Downloading [2.97 MB / 4.00 MB]

#> ⠼ Updating metadata [0/12] | Downloading [3.17 MB / 4.00 MB]

#> ⠴ Updating metadata [0/12] | Downloading [3.38 MB / 4.00 MB]

#> ⠦ Updating metadata [0/12] | Downloading [3.53 MB / 4.00 MB]

#> ⠧ Updating metadata [0/12] | Downloading [3.85 MB / 4.00 MB]

#> ✔ Updated metadata: 4.00 MB in 12 files.                    

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving standard (CRAN/BioC) packages

#>                                                                           

#> > pdi$solve()
#> > pdi$download()

#> ℹ Getting 43 pkgs (28.44 MB)

#> 

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Getting R6, assertthat,

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Getting R6, assertthat,

#>  1.4 MB/s ⸨             ⸩   4% |  0/43 pkg | ETA ~19.8s | Getting R6,          

#>  1.6 MB/s ⸨             ⸩   6% |  0/43 pkg | ETA ~17.3s | Getting callr, cli, 

#>  1.8 MB/s ⸨█            ⸩   8% |  0/43 pkg | ETA ~14.6s | Getting callr, cli, 

#>  1.8 MB/s ⸨█            ⸩   8% |  0/43 pkg | ETA ~14.7s | Getting callr,      

#>  1.8 MB/s ⸨█            ⸩   9% |  0/43 pkg | ETA ~14.7s | Getting callr, 

#>  1.6 MB/s ⸨█            ⸩  10% |  0/43 pkg | ETA ~16s   | Getting callr, 

#>  1.6 MB/s ⸨█            ⸩  10% |  0/43 pkg | ETA ~16.5s | Getting callr, 

#>  1.6 MB/s ⸨█            ⸩  11% |  0/43 pkg | ETA ~16.1s | Getting callr, 

#>  1.6 MB/s ⸨█            ⸩  12% |  0/43 pkg | ETA ~15.9s | Getting callr, 

#>  1.6 MB/s ⸨█            ⸩  13% |  0/43 pkg | ETA ~15.5s | Getting callr, 

#>  1.6 MB/s ⸨█            ⸩  14% |  0/43 pkg | ETA ~15s   | Getting callr, 

#>  1.6 MB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~15s   | Getting callr, 

#>                                                                          

#> ✔ Got backports 1.1.8 (macos) (71.78 kB)

#>  1.6 MB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~15s   | Getting callr, 

#>                                                                          

#> ✔ Got base64enc 0.1-3 (macos) (31.63 kB)

#>  1.6 MB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~15s   | Getting callr, 

#>                                                                          

#> ✔ Got desc 1.2.0 (macos) (285.96 kB)

#>  1.6 MB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~15s   | Getting callr, 

#>                                                                          

#> ✔ Got R6 2.4.1 (macos) (57.29 kB)

#>  1.6 MB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~15s   | Getting callr, 

#>                                                                          

#> ✔ Got assertthat 0.2.1 (macos) (52.39 kB)

#>  1.6 MB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~15s   | Getting callr, 

#>                                                                          

#> ✔ Got curl 4.3 (macos) (741.04 kB)

#>  1.6 MB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~15s   | Getting callr, 

#>                                                                          

#> ✔ Got cliapp 0.1.0 (macos) (232.97 kB)

#>  1.6 MB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~15s   | Getting callr, 

#>                                                                          

#> ✔ Got filelock 1.0.2 (macos) (26.68 kB)

#>  1.6 MB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~15s   | Getting callr, 

#>                                                                          

#> ✔ Got ellipsis 0.3.1 (macos) (33.34 kB)

#>  1.6 MB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~15s   | Getting callr, 

#>  1.5 MB/s ⸨██           ⸩  16% |  9/43 pkg | ETA ~15.5s | Getting callr, 

#>                                                                          

#> ✔ Got cli 2.0.2 (macos) (394.17 kB)

#>  1.5 MB/s ⸨██           ⸩  16% |  9/43 pkg | ETA ~15.5s | Getting callr, 

#>  1.6 MB/s ⸨██           ⸩  17% | 10/43 pkg | ETA ~14.9s | Getting callr, 

#>  1.6 MB/s ⸨██           ⸩  18% | 10/43 pkg | ETA ~14.2s | Getting callr, 

#>                                                                          

#> ✔ Got prettyunits 1.1.1 (macos) (34.63 kB)

#>  1.6 MB/s ⸨██           ⸩  18% | 10/43 pkg | ETA ~14.2s | Getting callr, 

#>  1.6 MB/s ⸨██           ⸩  19% | 11/43 pkg | ETA ~14.1s | Getting callr, 

#>  1.7 MB/s ⸨██           ⸩  20% | 11/43 pkg | ETA ~13.6s | Getting callr, 

#>  1.7 MB/s ⸨██           ⸩  21% | 11/43 pkg | ETA ~13.5s | Getting callr, 

#>                                                                          

#> ✔ Got rematch2 2.1.2 (macos) (44.78 kB)

#>  1.7 MB/s ⸨██           ⸩  21% | 11/43 pkg | ETA ~13.5s | Getting callr, 

#>  1.8 MB/s ⸨███          ⸩  23% | 12/43 pkg | ETA ~12.5s | Getting callr, 

#>                                                                          

#> ✔ Got lifecycle 0.2.0 (macos) (91.67 kB)

#>  1.8 MB/s ⸨███          ⸩  23% | 12/43 pkg | ETA ~12.5s | Getting callr, 

#>  1.8 MB/s ⸨███          ⸩  24% | 13/43 pkg | ETA ~12.1s | Getting callr, 

#>  1.9 MB/s ⸨███          ⸩  26% | 13/43 pkg | ETA ~11.3s | Getting callr, 

#>                                                                          

#> ✔ Got uuid 0.1-4 (macos) (27.76 kB)

#>  1.9 MB/s ⸨███          ⸩  26% | 13/43 pkg | ETA ~11.3s | Getting callr, 

#>                                                                          

#> ✔ Got pkgconfig 2.0.3 (macos) (17.74 kB)

#>  1.9 MB/s ⸨███          ⸩  26% | 13/43 pkg | ETA ~11.3s | Getting callr, 

#>                                                                          

#> ✔ Got hms 0.5.3 (macos) (103.23 kB)

#>  1.9 MB/s ⸨███          ⸩  26% | 13/43 pkg | ETA ~11.3s | Getting callr, 

#>  1.9 MB/s ⸨███          ⸩  28% | 16/43 pkg | ETA ~10.8s | Getting callr, 

#>                                                                          

#> ✔ Got progress 1.2.2 (macos) (81.76 kB)

#>  1.9 MB/s ⸨███          ⸩  28% | 16/43 pkg | ETA ~10.8s | Getting callr, 

#>  1.9 MB/s ⸨███          ⸩  29% | 17/43 pkg | ETA ~10.4s | Getting callr, 

#>                                                                          

#> ✔ Got magrittr 1.5 (macos) (152.67 kB)

#>  1.9 MB/s ⸨███          ⸩  29% | 17/43 pkg | ETA ~10.4s | Getting callr, 

#>  2.0 MB/s ⸨████         ⸩  31% | 18/43 pkg | ETA ~9.7s  | Getting callr, 

#>                                                                          

#> ✔ Got rprojroot 1.3-2 (macos) (78.69 kB)

#>  2.0 MB/s ⸨████         ⸩  31% | 18/43 pkg | ETA ~9.7s  | Getting callr, 

#>  2.1 MB/s ⸨████         ⸩  33% | 19/43 pkg | ETA ~9.2s  | Getting crayon, 

#>                                                                           

#> ✔ Got glue 1.4.1 (macos) (138.32 kB)

#>  2.1 MB/s ⸨████         ⸩  33% | 19/43 pkg | ETA ~9.2s  | Getting crayon, 

#>                                                                           

#> ✔ Got pkgbuild 1.0.8 (macos) (135.75 kB)

#>  2.1 MB/s ⸨████         ⸩  33% | 19/43 pkg | ETA ~9.2s  | Getting crayon, 

#>                                                                           

#> ✔ Got pillar 1.4.4 (macos) (179.45 kB)

#>  2.1 MB/s ⸨████         ⸩  33% | 19/43 pkg | ETA ~9.2s  | Getting crayon, 

#>  2.1 MB/s ⸨████         ⸩  35% | 22/43 pkg | ETA ~8.7s  | Getting crayon, 

#>                                                                           

#> ✔ Got rappdirs 0.3.1 (macos) (145.39 kB)

#>  2.1 MB/s ⸨████         ⸩  35% | 22/43 pkg | ETA ~8.7s  | Getting crayon, 

#>  2.3 MB/s ⸨████         ⸩  36% | 23/43 pkg | ETA ~7.7s  | Getting crayon, 

#>                                                                           

#> ✔ Got fansi 0.4.1 (macos) (210.34 kB)

#>  2.3 MB/s ⸨████         ⸩  36% | 23/43 pkg | ETA ~7.7s  | Getting crayon, 

#>  2.4 MB/s ⸨████         ⸩  38% | 24/43 pkg | ETA ~7.4s  | Getting crayon, 

#>                                                                           

#> ✔ Got digest 0.6.25 (macos) (245.72 kB)

#>  2.4 MB/s ⸨████         ⸩  38% | 24/43 pkg | ETA ~7.4s  | Getting crayon, 

#>                                                                           

#> ✔ Got prettycode 1.1.0 (macos) (242.89 kB)

#>  2.4 MB/s ⸨████         ⸩  38% | 24/43 pkg | ETA ~7.4s  | Getting crayon, 

#>                                                                           

#> ✔ Got processx 3.4.2 (macos) (252.45 kB)

#>  2.4 MB/s ⸨████         ⸩  38% | 24/43 pkg | ETA ~7.4s  | Getting crayon, 

#>                                                                           

#> ✔ Got stringr 1.4.0 (macos) (209.72 kB)

#>  2.4 MB/s ⸨████         ⸩  38% | 24/43 pkg | ETA ~7.4s  | Getting crayon, 

#>                                                                           

#> ✔ Got withr 2.2.0 (macos) (203.94 kB)

#>  2.4 MB/s ⸨████         ⸩  38% | 24/43 pkg | ETA ~7.4s  | Getting crayon, 

#>                                                                           

#> ✔ Got ps 1.3.3 (macos) (216.75 kB)

#>  2.4 MB/s ⸨████         ⸩  38% | 24/43 pkg | ETA ~7.4s  | Getting crayon, 

#>  2.4 MB/s ⸨█████        ⸩  41% | 30/43 pkg | ETA ~6.9s  | Getting crayon, 

#>                                                                           

#> ✔ Got utf8 1.1.4 (macos) (195.34 kB)

#>  2.4 MB/s ⸨█████        ⸩  41% | 30/43 pkg | ETA ~6.9s  | Getting crayon, 

#>  2.5 MB/s ⸨█████        ⸩  43% | 31/43 pkg | ETA ~6.5s  | Getting jsonlite, 

#>                                                                             

#> ✔ Got callr 3.4.3 (macos) (389.42 kB)

#>  2.5 MB/s ⸨█████        ⸩  43% | 31/43 pkg | ETA ~6.5s  | Getting jsonlite, 

#>  2.6 MB/s ⸨█████        ⸩  45% | 32/43 pkg | ETA ~6s    | Getting jsonlite, 

#>                                                                             

#> ✔ Got tibble 3.0.1 (macos) (387.43 kB)

#>  2.6 MB/s ⸨█████        ⸩  45% | 32/43 pkg | ETA ~6s    | Getting jsonlite, 

#>  2.7 MB/s ⸨██████       ⸩  48% | 33/43 pkg | ETA ~5.6s  | Getting jsonlite, 

#>  3.1 MB/s ⸨██████       ⸩  50% | 33/43 pkg | ETA ~4.6s  | Getting jsonlite, 

#>                                                                             

#> ✔ Got pak 0.1.2 (macos) (403.16 kB)

#>  3.1 MB/s ⸨██████       ⸩  50% | 33/43 pkg | ETA ~4.6s  | Getting jsonlite, 

#>                                                                             

#> ✔ Got selectr 0.4-2 (macos) (484.23 kB)

#>  3.1 MB/s ⸨██████       ⸩  50% | 33/43 pkg | ETA ~4.6s  | Getting jsonlite, 

#>  3.4 MB/s ⸨███████      ⸩  54% | 35/43 pkg | ETA ~3.9s  | Getting rlang,    

#>                                                                          

#> ✔ Got pkgcache 1.0.7 (macos) (523.24 kB)

#>  3.4 MB/s ⸨███████      ⸩  54% | 35/43 pkg | ETA ~3.9s  | Getting rlang, 

#>  3.5 MB/s ⸨███████      ⸩  57% | 36/43 pkg | ETA ~3.5s  | Getting stringi and 

#>  7.0 MB/s ⸨█████████████⸩ 100% | 36/43 pkg | ETA <1s    | Getting stringi and 

#>                                                                               

#> ✔ Got crayon 1.3.4 (macos) (748.32 kB)

#>  7.0 MB/s ⸨█████████████⸩ 100% | 36/43 pkg | ETA <1s    | Getting stringi and 

#>  6.8 MB/s ⸨█████████████⸩ 100% | 37/43 pkg | ETA <1s    | Getting stringi and 

#>                                                                               

#> ✔ Got lpSolve 5.6.15 (macos) (1.07 MB)

#>  6.8 MB/s ⸨█████████████⸩ 100% | 37/43 pkg | ETA <1s    | Getting stringi and 

#>  6.3 MB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi and 

#>                                                                               

#> ✔ Got vctrs 0.3.1 (macos) (1.24 MB)

#>  6.3 MB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi and 

#>                                                                               

#> ✔ Got jsonlite 1.7.0 (macos) (1.12 MB)

#>  6.3 MB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi and 

#>                                                                               

#> ✔ Got rlang 0.4.6 (macos) (1.18 MB)

#>  6.3 MB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi and 

#>                                                                               

#> ✔ Got xml2 1.3.2 (macos) (2.34 MB)

#>  6.3 MB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi and 

#>                                                                               

#> ✔ Got stringi 1.4.6 (macos) (13.62 MB)

#>  6.3 MB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi and 

#>  6.0 MB/s ⸨█████████████⸩ 100% | 43/43 pkg | ETA <1s    | Getting stringi and 

#>                                                                               

#> ✔ Downloaded 43 packages (28.44 MB) in 7.5s

#> > pdi$get_install_plan()
#> # A tibble: 43 x 42
#>    ref   type  direct directpkg status package version license needscompilation
#>    <chr> <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 R6    stan… FALSE  FALSE     OK     R6      2.4.1   MIT + … FALSE           
#>  2 asse… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  3 back… stan… FALSE  FALSE     OK     backpo… 1.1.8   GPL-2 … FALSE           
#>  4 base… stan… FALSE  FALSE     OK     base64… 0.1-3   GPL-2 … FALSE           
#>  5 callr stan… FALSE  FALSE     OK     callr   3.4.3   MIT + … FALSE           
#>  6 cli   stan… FALSE  FALSE     OK     cli     2.0.2   MIT + … FALSE           
#>  7 clia… stan… FALSE  FALSE     OK     cliapp  0.1.0   MIT + … FALSE           
#>  8 cray… stan… FALSE  FALSE     OK     crayon  1.3.4   MIT + … FALSE           
#>  9 curl  stan… FALSE  FALSE     OK     curl    4.3     MIT + … FALSE           
#> 10 desc  stan… FALSE  FALSE     OK     desc    1.2.0   MIT + … FALSE           
#> # … with 33 more rows, and 33 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repodir <chr>, target <chr>, deps <list>, mirror <chr>, sources <list>,
#> #   remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, cache_status <chr>, lib_status <chr>, old_version <chr>,
#> #   new_version <chr>, fulltarget <chr>, fulltarget_tree <chr>,
#> #   download_status <chr>, download_error <list>, file_size <dbl>,
#> #   library <chr>, binary <lgl>, dependencies <I<list>>, installed <chr>,
#> #   vignettes <lgl>, packaged <lgl>, file <chr>