Installation plans

An installation plan contains all data that is needed to install a set of package files. It is usually created from an installation proposal with solving the dependencies and downloading the package files.

Details

It is also possible to create an installation plan a different way. An installation plan object must be a data frame, with at least the following columns:

package: The name of the package.

type: The type of the package reference.

binary: Whether the package is a binary package.

file: Full path to the package file or directory.

dependencies: A list column that lists the names of the dependent packages for each package.

needscompilation: Whether the package needs compilation. This should be FALSE for binary packages.

For installation plans created via pkg_installation_proposal, the plan contains all columns from pkg_download_result objects, and some additional ones:

library: the library the package is supposed to be installed to.

direct: whether the package was directly requested or it is installed as a dependency.

vignettes: whether the vignettes need to be (re)built.

packaged: whether R CMD build was already called for the package.

See also

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file7596eb0987b

pdi$resolve()

#> 

#> ✔ Updated metadata: 4.17 MB in 12 files.

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving standard (CRAN/BioC) packages

#>                                                                           

pdi$solve()
pdi$download()

#> ℹ Getting 43 pkgs (28.76 MB)

#> 

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Getting R6, assertthat,

#>           ⸨             ⸩   0% |  0/43 pkg | ETA ??s  | Getting R6, assertthat,

#>  1.0 MB/s ⸨             ⸩   4% |  0/43 pkg | ETA ~26.5s | Getting R6,          

#>  1.4 MB/s ⸨             ⸩   7% |  0/43 pkg | ETA ~19.2s | Getting callr, 

#>  1.6 MB/s ⸨█            ⸩   9% |  0/43 pkg | ETA ~16.9s | Getting cliapp, 

#>  1.6 MB/s ⸨█            ⸩  10% |  0/43 pkg | ETA ~16s   | Getting cliapp, 

#>  1.6 MB/s ⸨█            ⸩  11% |  0/43 pkg | ETA ~15.9s | Getting cliapp, 

#>  1.6 MB/s ⸨█            ⸩  12% |  0/43 pkg | ETA ~15.9s | Getting cliapp, 

#>  1.7 MB/s ⸨█            ⸩  14% |  0/43 pkg | ETA ~14.4s | Getting cliapp, 

#>  1.4 MB/s ⸨█            ⸩  14% |  0/43 pkg | ETA ~17.5s | Getting cliapp, 

#>  969 kB/s ⸨█            ⸩  15% |  0/43 pkg | ETA ~25.2s | Getting cliapp, 

#>  186 kB/s ⸨██           ⸩  17% |  0/43 pkg | ETA ~2m    | Getting cliapp, 

#>  164 kB/s ⸨██           ⸩  18% |  0/43 pkg | ETA ~2m    | Getting cliapp, 

#>  139 kB/s ⸨██           ⸩  19% |  0/43 pkg | ETA ~2m    | Getting cliapp, 

#>  159 kB/s ⸨██           ⸩  21% |  0/43 pkg | ETA ~2m    | Getting cliapp, 

#>  133 kB/s ⸨██           ⸩  21% |  0/43 pkg | ETA ~2m    | Getting cliapp, 

#>                                                                           

#> ✔ Got R6 2.5.0 (macos) (82.45 kB)

#>  133 kB/s ⸨██           ⸩  21% |  0/43 pkg | ETA ~2m    | Getting cliapp, 

#>                                                                           

#> ✔ Got assertthat 0.2.1 (macos) (52.57 kB)

#>  133 kB/s ⸨██           ⸩  21% |  0/43 pkg | ETA ~2m    | Getting cliapp, 

#>                                                                           

#> ✔ Got backports 1.2.0 (macos) (86.17 kB)

#>  133 kB/s ⸨██           ⸩  21% |  0/43 pkg | ETA ~2m    | Getting cliapp, 

#>                                                                           

#> ✔ Got base64enc 0.1-3 (macos) (31.70 kB)

#>  133 kB/s ⸨██           ⸩  21% |  0/43 pkg | ETA ~2m    | Getting cliapp, 

#>                                                                           

#> ✔ Got cli 2.1.0 (macos) (406.23 kB)

#>  133 kB/s ⸨██           ⸩  21% |  0/43 pkg | ETA ~2m    | Getting cliapp, 

#>  123 kB/s ⸨██           ⸩  23% |  5/43 pkg | ETA ~3m    | Getting cliapp, 

#>                                                                           

#> ✔ Got callr 3.5.1 (macos) (385.44 kB)

#>  123 kB/s ⸨██           ⸩  23% |  5/43 pkg | ETA ~3m    | Getting cliapp, 

#>  187 kB/s ⸨███          ⸩  24% |  6/43 pkg | ETA ~1m    | Getting cliapp, 

#>                                                                           

#> ✔ Got ellipsis 0.3.1 (macos) (33.50 kB)

#>  187 kB/s ⸨███          ⸩  24% |  6/43 pkg | ETA ~1m    | Getting cliapp, 

#>                                                                           

#> ✔ Got pkgconfig 2.0.3 (macos) (17.74 kB)

#>  187 kB/s ⸨███          ⸩  24% |  6/43 pkg | ETA ~1m    | Getting cliapp, 

#>  322 kB/s ⸨███          ⸩  26% |  8/43 pkg | ETA ~1m    | Getting cliapp, 

#>                                                                           

#> ✔ Got filelock 1.0.2 (macos) (26.73 kB)

#>  322 kB/s ⸨███          ⸩  26% |  8/43 pkg | ETA ~1m    | Getting cliapp, 

#>  302 kB/s ⸨███          ⸩  27% |  9/43 pkg | ETA ~1m    | Getting crayon, desc,

#>  343 kB/s ⸨███          ⸩  28% |  9/43 pkg | ETA ~1m    | Getting crayon, desc,

#>                                                                                

#> ✔ Got curl 4.3 (macos) (741.52 kB)

#>  343 kB/s ⸨███          ⸩  28% |  9/43 pkg | ETA ~1m    | Getting crayon, desc,

#>                                                                                

#> ✔ Got rematch2 2.1.2 (macos) (44.81 kB)

#>  343 kB/s ⸨███          ⸩  28% |  9/43 pkg | ETA ~1m    | Getting crayon, desc,

#>  414 kB/s ⸨███          ⸩  29% | 11/43 pkg | ETA ~49s   | Getting crayon, desc,

#>                                                                                

#> ✔ Got prettyunits 1.1.1 (macos) (34.70 kB)

#>  414 kB/s ⸨███          ⸩  29% | 11/43 pkg | ETA ~49s   | Getting crayon, desc,

#>                                                                                

#> ✔ Got uuid 0.1-4 (macos) (27.81 kB)

#>  414 kB/s ⸨███          ⸩  29% | 11/43 pkg | ETA ~49s   | Getting crayon, desc,

#>  366 kB/s ⸨███          ⸩  30% | 13/43 pkg | ETA ~55s   | Getting crayon,      

#>  338 kB/s ⸨███          ⸩  31% | 13/43 pkg | ETA ~59.1s | Getting crayon, 

#>                                                                           

#> ✔ Got lifecycle 0.2.0 (macos) (91.60 kB)

#>  338 kB/s ⸨███          ⸩  31% | 13/43 pkg | ETA ~59.1s | Getting crayon, 

#>                                                                           

#> ✔ Got hms 0.5.3 (macos) (103.86 kB)

#>  338 kB/s ⸨███          ⸩  31% | 13/43 pkg | ETA ~59.1s | Getting crayon, 

#>  499 kB/s ⸨████         ⸩  33% | 15/43 pkg | ETA ~38.6s | Getting crayon, 

#>                                                                           

#> ✔ Got glue 1.4.2 (macos) (139.02 kB)

#>  499 kB/s ⸨████         ⸩  33% | 15/43 pkg | ETA ~38.6s | Getting crayon, 

#>                                                                           

#> ✔ Got progress 1.2.2 (macos) (83.46 kB)

#>  499 kB/s ⸨████         ⸩  33% | 15/43 pkg | ETA ~38.6s | Getting crayon, 

#>  695 kB/s ⸨████         ⸩  35% | 17/43 pkg | ETA ~26.8s | Getting crayon, 

#>                                                                           

#> ✔ Got rprojroot 1.3-2 (macos) (79.15 kB)

#>  695 kB/s ⸨████         ⸩  35% | 17/43 pkg | ETA ~26.8s | Getting crayon, 

#>  736 kB/s ⸨████         ⸩  36% | 18/43 pkg | ETA ~25s   | Getting crayon, 

#>                                                                           

#> ✔ Got magrittr 1.5 (macos) (153.23 kB)

#>  736 kB/s ⸨████         ⸩  36% | 18/43 pkg | ETA ~25s   | Getting crayon, 

#>                                                                           

#> ✔ Got pillar 1.4.6 (macos) (176.43 kB)

#>  736 kB/s ⸨████         ⸩  36% | 18/43 pkg | ETA ~25s   | Getting crayon, 

#>                                                                           

#> ✔ Got pkgbuild 1.1.0 (macos) (139.36 kB)

#>  736 kB/s ⸨████         ⸩  36% | 18/43 pkg | ETA ~25s   | Getting crayon, 

#>                                                                           

#> ✔ Got rappdirs 0.3.1 (macos) (145.34 kB)

#>  736 kB/s ⸨████         ⸩  36% | 18/43 pkg | ETA ~25s   | Getting crayon, 

#>  834 kB/s ⸨████         ⸩  38% | 22/43 pkg | ETA ~21.6s | Getting crayon, 

#>  717 kB/s ⸨████         ⸩  38% | 22/43 pkg | ETA ~24.9s | Getting crayon, 

#>                                                                           

#> ✔ Got cliapp 0.1.1 (macos) (235.74 kB)

#>  717 kB/s ⸨████         ⸩  38% | 22/43 pkg | ETA ~24.9s | Getting crayon, 

#>                                                                           

#> ✔ Got fansi 0.4.1 (macos) (210.70 kB)

#>  717 kB/s ⸨████         ⸩  38% | 22/43 pkg | ETA ~24.9s | Getting crayon, 

#>                                                                           

#> ✔ Got utf8 1.1.4 (macos) (195.53 kB)

#>  717 kB/s ⸨████         ⸩  38% | 22/43 pkg | ETA ~24.9s | Getting crayon, 

#>  884 kB/s ⸨█████        ⸩  40% | 25/43 pkg | ETA ~19.4s | Getting crayon, 

#>                                                                           

#> ✔ Got prettycode 1.1.0 (macos) (242.92 kB)

#>  884 kB/s ⸨█████        ⸩  40% | 25/43 pkg | ETA ~19.4s | Getting crayon, 

#>                                                                           

#> ✔ Got ps 1.4.0 (macos) (251.68 kB)

#>  884 kB/s ⸨█████        ⸩  40% | 25/43 pkg | ETA ~19.4s | Getting crayon, 

#>  754 kB/s ⸨█████        ⸩  41% | 27/43 pkg | ETA ~22.5s | Getting crayon, 

#>                                                                           

#> ✔ Got withr 2.3.0 (macos) (205.49 kB)

#>  754 kB/s ⸨█████        ⸩  41% | 27/43 pkg | ETA ~22.5s | Getting crayon, 

#>                                                                           

#> ✔ Got digest 0.6.27 (macos) (300.37 kB)

#>  754 kB/s ⸨█████        ⸩  41% | 27/43 pkg | ETA ~22.5s | Getting crayon, 

#>                                                                           

#> ✔ Got desc 1.2.0 (macos) (285.23 kB)

#>  754 kB/s ⸨█████        ⸩  41% | 27/43 pkg | ETA ~22.5s | Getting crayon, 

#>                                                                           

#> ✔ Got stringr 1.4.0 (macos) (210.65 kB)

#>  754 kB/s ⸨█████        ⸩  41% | 27/43 pkg | ETA ~22.5s | Getting crayon, 

#>  932 kB/s ⸨█████        ⸩  43% | 31/43 pkg | ETA ~17.4s | Getting crayon, 

#>                                                                           

#> ✔ Got processx 3.4.4 (macos) (278.50 kB)

#>  932 kB/s ⸨█████        ⸩  43% | 31/43 pkg | ETA ~17.4s | Getting crayon, 

#>  979 kB/s ⸨█████        ⸩  46% | 32/43 pkg | ETA ~16s   | Getting crayon, 

#>  1.2 MB/s ⸨██████       ⸩  49% | 32/43 pkg | ETA ~12.4s | Getting crayon, 

#>  1.5 MB/s ⸨███████      ⸩  54% | 32/43 pkg | ETA ~9.1s  | Getting stringi, 

#>  1.8 MB/s ⸨███████      ⸩  58% | 32/43 pkg | ETA ~6.6s  | Getting stringi and 

#>                                                                               

#> ✔ Got tibble 3.0.4 (macos) (393.07 kB)

#>  1.8 MB/s ⸨███████      ⸩  58% | 32/43 pkg | ETA ~6.6s  | Getting stringi and 

#>  2.4 MB/s ⸨████████     ⸩  66% | 33/43 pkg | ETA ~4s    | Getting stringi and 

#>  3.3 MB/s ⸨██████████   ⸩  78% | 33/43 pkg | ETA ~2s    | Getting stringi and 

#>                                                                               

#> ✔ Got pak 0.1.2 (macos) (403.75 kB)

#>  3.3 MB/s ⸨██████████   ⸩  78% | 33/43 pkg | ETA ~2s    | Getting stringi and 

#>  4.2 MB/s ⸨███████████  ⸩  91% | 34/43 pkg | ETA <1s    | Getting stringi and 

#>                                                                               

#> ✔ Got selectr 0.4-2 (macos) (485.31 kB)

#>  4.2 MB/s ⸨███████████  ⸩  91% | 34/43 pkg | ETA <1s    | Getting stringi and 

#>  685 kB/s ⸨█████████████⸩ 100% | 35/43 pkg | ETA <1s    | Getting stringi     

#>                                                                          

#> ✔ Got pkgcache 1.1.1 (macos) (551.45 kB)

#>  685 kB/s ⸨█████████████⸩ 100% | 35/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got jsonlite 1.7.1 (macos) (1.12 MB)

#>  685 kB/s ⸨█████████████⸩ 100% | 35/43 pkg | ETA <1s    | Getting stringi

#>  669 kB/s ⸨█████████████⸩ 100% | 37/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got lpSolve 5.6.15 (macos) (1.07 MB)

#>  669 kB/s ⸨█████████████⸩ 100% | 37/43 pkg | ETA <1s    | Getting stringi

#>  693 kB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got crayon 1.3.4 (macos) (748.90 kB)

#>  693 kB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got rlang 0.4.8 (macos) (1.21 MB)

#>  693 kB/s ⸨█████████████⸩ 100% | 38/43 pkg | ETA <1s    | Getting stringi

#>  685 kB/s ⸨█████████████⸩ 100% | 40/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got xml2 1.3.2 (macos) (2.34 MB)

#>  685 kB/s ⸨█████████████⸩ 100% | 40/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got vctrs 0.3.4 (macos) (1.31 MB)

#>  685 kB/s ⸨█████████████⸩ 100% | 40/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Got stringi 1.5.3 (macos) (13.64 MB)

#>  685 kB/s ⸨█████████████⸩ 100% | 40/43 pkg | ETA <1s    | Getting stringi

#>  668 kB/s ⸨█████████████⸩ 100% | 43/43 pkg | ETA <1s    | Getting stringi

#>                                                                          

#> ✔ Downloaded 43 packages (28.77 MB) in 37.8s

pdi$get_install_plan()

#> # A tibble: 43 x 42
#>    ref   type  direct directpkg status package version license needscompilation
#>    <chr> <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 R6    stan… FALSE  FALSE     OK     R6      2.5.0   MIT + … FALSE           
#>  2 asse… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  3 back… stan… FALSE  FALSE     OK     backpo… 1.2.0   GPL-2 … FALSE           
#>  4 base… stan… FALSE  FALSE     OK     base64… 0.1-3   GPL-2 … FALSE           
#>  5 callr stan… FALSE  FALSE     OK     callr   3.5.1   MIT + … FALSE           
#>  6 cli   stan… FALSE  FALSE     OK     cli     2.1.0   MIT + … FALSE           
#>  7 clia… stan… FALSE  FALSE     OK     cliapp  0.1.1   MIT + … FALSE           
#>  8 cray… stan… FALSE  FALSE     OK     crayon  1.3.4   MIT + … FALSE           
#>  9 curl  stan… FALSE  FALSE     OK     curl    4.3     MIT + … FALSE           
#> 10 desc  stan… FALSE  FALSE     OK     desc    1.2.0   MIT + … FALSE           
#> # … with 33 more rows, and 33 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repodir <chr>, target <chr>, deps <list>, mirror <chr>, sources <list>,
#> #   remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, cache_status <chr>, lib_status <chr>, old_version <chr>,
#> #   new_version <chr>, fulltarget <chr>, fulltarget_tree <chr>,
#> #   download_status <chr>, download_error <list>, file_size <dbl>,
#> #   library <chr>, binary <lgl>, dependencies <I<list>>, installed <chr>,
#> #   vignettes <lgl>, packaged <lgl>, file <chr>