Installation plans

An installation plan contains all data that is needed to install a set of package files. It is usually created from an installation proposal with solving the dependencies and downloading the package files.

Details

It is also possible to create an installation plan a different way. An installation plan object must be a data frame, with at least the following columns:

package: The name of the package.

type: The type of the package reference.

binary: Whether the package is a binary package.

file: Full path to the package file or directory.

dependencies: A list column that lists the names of the dependent packages for each package.

needscompilation: Whether the package needs compilation. This should be FALSE for binary packages.

For installation plans created via pkg_installation_proposal, the plan contains all columns from pkg_download_result objects, and some additional ones:

library: the library the package is supposed to be installed to.

direct: whether the package was directly requested or it is installed as a dependency.

vignettes: whether the vignettes need to be (re)built.

packaged: whether R CMD build was already called for the package.

See also

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpSQ9qwg/file7d56a19086d

pdi$resolve()

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving standard (CRAN/BioC) packages

#>  Found  1  deps for  0/1  pkgs [⠙] Resolving standard (CRAN/BioC) packages

#>  Found  1  deps for  0/1  pkgs [⠹] Resolving standard (CRAN/BioC) packages

#>  Found  1  deps for  0/1  pkgs [⠸] Resolving standard (CRAN/BioC) packages

#>  Found  1  deps for  0/1  pkgs [⠼] Resolving standard (CRAN/BioC) packages

#> ✔ Updated metadata: 4.21 MB in 12 files.

#> 

#>  Found  1  deps for  0/1  pkgs [⠼] Resolving standard (CRAN/BioC) packages

#> ℹ Updating local metadata database

#> ✔ Updating local metadata database ... done

#> 

#>  Found  1  deps for  0/1  pkgs [⠼] Resolving standard (CRAN/BioC) packages

#>  Found  1  deps for  0/1  pkgs [⠴] Resolving standard (CRAN/BioC) packages

#>  Found  1  deps for  0/1  pkgs [⠦] Resolving standard (CRAN/BioC) packages

#>  Found  1  deps for  0/1  pkgs [⠧] Resolving standard (CRAN/BioC) packages

pdi$solve()
pdi$download()

#> ℹ Getting 39 pkgs (26.83 MB) and 3 pkgs with unknown sizes

#> 

#>           ⸨             ⸩   0% |  0/42 pkg | Connecting...

#>           ⸨             ⸩   0% |  0/42 pkg | Connecting...

#>           ⸨             ⸩   0% |  0/42 pkg | Getting R6, assertthat, base64enc,

#>           ⸨             ⸩   0% |  0/42 pkg | Getting R6, assertthat, base64enc,

#>  2.2 MB/s ⸨             ⸩   0% |  0/42 pkg | Getting R6, assertthat, base64enc,

#>  2.4 MB/s ⸨█            ⸩   8% |  0/42 pkg | ETA ~10.9s | Getting cliapp, 

#>  2.3 MB/s ⸨█            ⸩   9% |  0/42 pkg | ETA ~11.1s | Getting cliapp, 

#>  2.4 MB/s ⸨█            ⸩  10% |  0/42 pkg | ETA ~10.9s | Getting cliapp, 

#>  2.2 MB/s ⸨█            ⸩  10% |  0/42 pkg | ETA ~11.7s | Getting cliapp, 

#>  2.0 MB/s ⸨█            ⸩  11% |  0/42 pkg | ETA ~12.7s | Getting cliapp, 

#>  2.0 MB/s ⸨█            ⸩  12% |  0/42 pkg | ETA ~12.6s | Getting cliapp, 

#>  1.9 MB/s ⸨█            ⸩  13% |  0/42 pkg | ETA ~13s   | Getting cliapp, 

#>  1.9 MB/s ⸨█            ⸩  13% |  0/42 pkg | ETA ~12.7s | Getting cliapp, 

#>  1.9 MB/s ⸨█            ⸩  14% |  0/42 pkg | ETA ~12.9s | Getting cliapp, 

#> ✔ Got assertthat 0.2.1 (macos) (52.57 kB)

#>  1.9 MB/s ⸨█            ⸩  14% |  0/42 pkg | ETA ~12.9s | Getting cliapp, 

#> ✔ Got base64enc 0.1-3 (macos) (31.70 kB)

#>  1.9 MB/s ⸨█            ⸩  14% |  0/42 pkg | ETA ~12.9s | Getting cliapp, 

#> ✔ Got callr 3.5.1 (macos) (385.44 kB)

#>  1.9 MB/s ⸨█            ⸩  14% |  0/42 pkg | ETA ~12.9s | Getting cliapp, 

#> ✔ Got cli 2.1.0 (macos) (406.23 kB)

#>  1.9 MB/s ⸨█            ⸩  14% |  0/42 pkg | ETA ~12.9s | Getting cliapp, 

#> ✔ Got desc 1.2.0 (macos) (285.23 kB)

#>  1.9 MB/s ⸨█            ⸩  14% |  0/42 pkg | ETA ~12.9s | Getting cliapp, 

#> ✔ Got filelock 1.0.2 (macos) (26.73 kB)

#>  1.9 MB/s ⸨█            ⸩  14% |  0/42 pkg | ETA ~12.9s | Getting cliapp, 

#> ✔ Got R6 2.5.0 (macos) (82.45 kB)

#>  1.9 MB/s ⸨█            ⸩  14% |  0/42 pkg | ETA ~12.9s | Getting cliapp, 

#>  1.9 MB/s ⸨██           ⸩  16% |  7/42 pkg | ETA ~12.8s | Getting cliapp, 

#> ✔ Got curl 4.3 (macos) (741.52 kB)

#>  1.9 MB/s ⸨██           ⸩  16% |  7/42 pkg | ETA ~12.8s | Getting cliapp, 

#> ✔ Got ellipsis 0.3.1 (macos) (33.50 kB)

#>  1.9 MB/s ⸨██           ⸩  16% |  7/42 pkg | ETA ~12.8s | Getting cliapp, 

#>  2.0 MB/s ⸨██           ⸩  18% |  9/42 pkg | ETA ~11.8s | Getting cliapp, 

#>  2.0 MB/s ⸨██           ⸩  20% |  9/42 pkg | ETA ~11.3s | Getting cliapp, 

#> ✔ Got prettyunits 1.1.1 (macos) (34.70 kB)

#>  2.0 MB/s ⸨██           ⸩  20% |  9/42 pkg | ETA ~11.3s | Getting cliapp, 

#>  2.1 MB/s ⸨██           ⸩  22% | 10/42 pkg | ETA ~10.4s | Getting cliapp, 

#>  2.1 MB/s ⸨██           ⸩  23% | 10/42 pkg | ETA ~10.3s | Getting cliapp, 

#>  2.2 MB/s ⸨███          ⸩  25% | 10/42 pkg | ETA ~9.6s  | Getting cliapp, 

#>  2.3 MB/s ⸨███          ⸩  27% | 10/42 pkg | ETA ~9s    | Getting cliapp, 

#> ✔ Got rematch2 2.1.2 (macos) (44.81 kB)

#>  2.3 MB/s ⸨███          ⸩  27% | 10/42 pkg | ETA ~9s    | Getting cliapp, 

#> ✔ Got pkgconfig 2.0.3 (macos) (17.74 kB)

#>  2.3 MB/s ⸨███          ⸩  27% | 10/42 pkg | ETA ~9s    | Getting cliapp, 

#> ✔ Got uuid 0.1-4 (macos) (27.81 kB)

#>  2.3 MB/s ⸨███          ⸩  27% | 10/42 pkg | ETA ~9s    | Getting cliapp, 

#>  2.4 MB/s ⸨███          ⸩  29% | 13/42 pkg | ETA ~8.4s  | Getting crayon, 

#>  2.5 MB/s ⸨███          ⸩  31% | 13/42 pkg | ETA ~8.1s  | Getting crayon, 

#> ✔ Got hms 0.5.3 (macos) (103.86 kB)

#>  2.5 MB/s ⸨███          ⸩  31% | 13/42 pkg | ETA ~8.1s  | Getting crayon, 

#>  2.5 MB/s ⸨████         ⸩  33% | 14/42 pkg | ETA ~7.6s  | Getting crayon, 

#>  2.6 MB/s ⸨████         ⸩  34% | 14/42 pkg | ETA ~7.3s  | Getting crayon, 

#> ✔ Got lifecycle 0.2.0 (macos) (91.60 kB)

#>  2.6 MB/s ⸨████         ⸩  34% | 14/42 pkg | ETA ~7.3s  | Getting crayon, 

#> ✔ Got progress 1.2.2 (macos) (83.46 kB)

#>  2.6 MB/s ⸨████         ⸩  34% | 14/42 pkg | ETA ~7.3s  | Getting crayon, 

#> ✔ Got rprojroot 2.0.2 (macos) (96.29 kB)

#>  2.6 MB/s ⸨████         ⸩  34% | 14/42 pkg | ETA ~7.3s  | Getting crayon, 

#>  2.6 MB/s ⸨████         ⸩  36% | 17/42 pkg | ETA ~7s    | Getting crayon, 

#>  2.7 MB/s ⸨████         ⸩  37% | 17/42 pkg | ETA ~6.6s  | Getting crayon, 

#> ✔ Got glue 1.4.2 (macos) (139.02 kB)

#>  2.7 MB/s ⸨████         ⸩  37% | 17/42 pkg | ETA ~6.6s  | Getting crayon, 

#>  2.9 MB/s ⸨█████        ⸩  40% | 18/42 pkg | ETA ~6s    | Getting jsonlite, 

#> ✔ Got rappdirs 0.3.1 (macos) (145.34 kB)

#>  2.9 MB/s ⸨█████        ⸩  40% | 18/42 pkg | ETA ~6s    | Getting jsonlite, 

#> ✔ Got pillar 1.4.6 (macos) (176.43 kB)

#>  2.9 MB/s ⸨█████        ⸩  40% | 18/42 pkg | ETA ~6s    | Getting jsonlite, 

#> ✔ Got pkgbuild 1.1.0 (macos) (139.36 kB)

#>  2.9 MB/s ⸨█████        ⸩  40% | 18/42 pkg | ETA ~6s    | Getting jsonlite, 

#> ✔ Got utf8 1.1.4 (macos) (195.53 kB)

#>  2.9 MB/s ⸨█████        ⸩  40% | 18/42 pkg | ETA ~6s    | Getting jsonlite, 

#> ✔ Got pak 0.1.2.1 (source) (159.96 kB)

#>  2.9 MB/s ⸨█████        ⸩  40% | 18/42 pkg | ETA ~6s    | Getting jsonlite, 

#>  3.1 MB/s ⸨█████        ⸩  44% | 23/42 pkg | ETA ~5.2s  | Getting jsonlite, 

#> ✔ Got cliapp 0.1.1 (macos) (235.74 kB)

#>  3.1 MB/s ⸨█████        ⸩  44% | 23/42 pkg | ETA ~5.2s  | Getting jsonlite, 

#> ✔ Got magrittr 2.0.1 (macos) (224.85 kB)

#>  3.1 MB/s ⸨█████        ⸩  44% | 23/42 pkg | ETA ~5.2s  | Getting jsonlite, 

#> ✔ Got fansi 0.4.1 (macos) (210.70 kB)

#>  3.1 MB/s ⸨█████        ⸩  44% | 23/42 pkg | ETA ~5.2s  | Getting jsonlite, 

#> ✔ Got prettycode 1.1.0 (macos) (242.92 kB)

#>  3.1 MB/s ⸨█████        ⸩  44% | 23/42 pkg | ETA ~5.2s  | Getting jsonlite, 

#> ✔ Got ps 1.4.0 (macos) (251.68 kB)

#>  3.1 MB/s ⸨█████        ⸩  44% | 23/42 pkg | ETA ~5.2s  | Getting jsonlite, 

#>  3.3 MB/s ⸨██████       ⸩  48% | 28/42 pkg | ETA ~4.5s  | Getting jsonlite, 

#> ✔ Got stringr 1.4.0 (macos) (210.65 kB)

#>  3.3 MB/s ⸨██████       ⸩  48% | 28/42 pkg | ETA ~4.5s  | Getting jsonlite, 

#> ✔ Got withr 2.3.0 (macos) (205.49 kB)

#>  3.3 MB/s ⸨██████       ⸩  48% | 28/42 pkg | ETA ~4.5s  | Getting jsonlite, 

#> ✔ Got digest 0.6.27 (macos) (300.37 kB)

#>  3.3 MB/s ⸨██████       ⸩  48% | 28/42 pkg | ETA ~4.5s  | Getting jsonlite, 

#> ✔ Got processx 3.4.4 (macos) (278.50 kB)

#>  3.3 MB/s ⸨██████       ⸩  48% | 28/42 pkg | ETA ~4.5s  | Getting jsonlite, 

#>  3.4 MB/s ⸨██████       ⸩  52% | 32/42 pkg | ETA ~4s    | Getting jsonlite, 

#>  3.6 MB/s ⸨███████      ⸩  55% | 32/42 pkg | ETA ~3.6s  | Getting jsonlite, 

#>  3.9 MB/s ⸨███████      ⸩  61% | 32/42 pkg | ETA ~2.8s  | Getting jsonlite, 

#> ✔ Got tibble 3.0.4 (macos) (393.07 kB)

#>  3.9 MB/s ⸨███████      ⸩  61% | 32/42 pkg | ETA ~2.8s  | Getting jsonlite, 

#> ✔ Got crayon 1.3.4 (macos) (748.90 kB)

#>  3.9 MB/s ⸨███████      ⸩  61% | 32/42 pkg | ETA ~2.8s  | Getting jsonlite, 

#>  4.4 MB/s ⸨█████████    ⸩  69% | 34/42 pkg | ETA ~2s    | Getting stringi, 

#> ✔ Got selectr 0.4-2 (macos) (485.31 kB)

#>  4.4 MB/s ⸨█████████    ⸩  69% | 34/42 pkg | ETA ~2s    | Getting stringi, 

#> ✔ Got pkgcache 1.1.1 (macos) (551.45 kB)

#>  4.4 MB/s ⸨█████████    ⸩  69% | 34/42 pkg | ETA ~2s    | Getting stringi, 

#>  5.9 MB/s ⸨██████████   ⸩  78% | 36/42 pkg | ETA ~1.1s  | Getting stringi and 

#> ✔ Got lpSolve 5.6.15 (macos) (1.07 MB)

#>  5.9 MB/s ⸨██████████   ⸩  78% | 36/42 pkg | ETA ~1.1s  | Getting stringi and 

#>  6.4 MB/s ⸨███████████  ⸩  87% | 37/42 pkg | ETA <1s    | Getting stringi and 

#>  6.8 MB/s ⸨████████████ ⸩  95% | 37/42 pkg | ETA <1s    | Getting stringi and 

#> ✔ Got jsonlite 1.7.1 (macos) (1.12 MB)

#>  6.8 MB/s ⸨████████████ ⸩  95% | 37/42 pkg | ETA <1s    | Getting stringi and 

#> ✔ Got rlang 0.4.8 (macos) (1.21 MB)

#>  6.8 MB/s ⸨████████████ ⸩  95% | 37/42 pkg | ETA <1s    | Getting stringi and 

#>  7.0 MB/s ⸨█████████████⸩ 100% | 39/42 pkg | ETA <1s    | Getting stringi and 

#> ✔ Got vctrs 0.3.5 (macos) (1.40 MB)

#>  7.0 MB/s ⸨█████████████⸩ 100% | 39/42 pkg | ETA <1s    | Getting stringi and 

#>  6.6 MB/s ⸨█████████████⸩ 100% | 40/42 pkg | ETA <1s    | Getting stringi and 

#> ✔ Got xml2 1.3.2 (macos) (2.34 MB)

#>  6.6 MB/s ⸨█████████████⸩ 100% | 40/42 pkg | ETA <1s    | Getting stringi and 

#> ✔ Got stringi 1.5.3 (macos) (13.64 MB)

#>  6.6 MB/s ⸨█████████████⸩ 100% | 40/42 pkg | ETA <1s    | Getting stringi and 

#> ✔ Downloaded 42 packages (28.62 MB) in 5.8s

pdi$get_install_plan()

#> # A tibble: 42 x 42
#>    ref   type  direct directpkg status package version license needscompilation
#>    <chr> <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 R6    stan… FALSE  FALSE     OK     R6      2.5.0   MIT + … FALSE           
#>  2 asse… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  3 base… stan… FALSE  FALSE     OK     base64… 0.1-3   GPL-2 … FALSE           
#>  4 callr stan… FALSE  FALSE     OK     callr   3.5.1   MIT + … FALSE           
#>  5 cli   stan… FALSE  FALSE     OK     cli     2.1.0   MIT + … FALSE           
#>  6 clia… stan… FALSE  FALSE     OK     cliapp  0.1.1   MIT + … FALSE           
#>  7 cray… stan… FALSE  FALSE     OK     crayon  1.3.4   MIT + … FALSE           
#>  8 curl  stan… FALSE  FALSE     OK     curl    4.3     MIT + … FALSE           
#>  9 desc  stan… FALSE  FALSE     OK     desc    1.2.0   MIT + … FALSE           
#> 10 dige… stan… FALSE  FALSE     OK     digest  0.6.27  GPL (>… FALSE           
#> # … with 32 more rows, and 33 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repodir <chr>, target <chr>, deps <list>, mirror <chr>, sources <list>,
#> #   remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, cache_status <chr>, lib_status <chr>, old_version <chr>,
#> #   new_version <chr>, fulltarget <chr>, fulltarget_tree <chr>,
#> #   download_status <chr>, download_error <list>, file_size <dbl>,
#> #   library <chr>, binary <lgl>, dependencies <I<list>>, installed <chr>,
#> #   vignettes <lgl>, packaged <lgl>, file <chr>