Configuration

Configuration options for several pkgdepends classes. Not all classes use all of these options. E.g. a pkg_download_proposal is not concerned about package libraries, so it'll ignore the library option.

Details

Options:

library: package library to use for checking already installed packages when considering dependencies in dependency lookup or package installation. Defaults to the first path in .libPaths().

cache_dir: directory to download the packages to. Defaults to a temporary directory within the R session temporary directory, see base::tempdir().

package_cache_dir: package cache location of pkgcache::package_cache. The default is the pkgcache default.

metadata_cache_dir: location of metadata replica of pkgcache::cranlike_metadata_cache. Defaults to a temporary directory within the R session temporary directory, see base::tempdir().

platforms: Character vector of platforms to download or install for. Possible platforms are windows, macos and source. Defaults to the current platform, and source.

cran-mirror: CRAN mirror to use. Defaults to the repos option (see base::options()), if that's not set then https://cran.rstudio.com.

dependencies: Dependencies to consider or download or install. Defaults to the hard dependencies, see pkg_dep_types_hard().

r-versions: Character vector, R versions to download or install packages for. It defaults to the current R version.

build-vignettes: Whether to build vignettes for package trees. This is only used if the package is obtained from a package tree, and not from a source (or binary) package archive. By default vignettes are not built in this case. If you set this to TRUE, then you need to make sure that the vignette builder packages are available, as these are not installed by default currently.

metadata-update-after: A time interval as a difftime object. pkgdepends will update the metadata cache if it is older than this. The default is one day.