R6 class for package dependency lookup

Look up dependencies of R packages from various sources.

new_pkg_deps(refs, ...)

Arguments

refs
Package names or references. See 'Package references' for the syntax.

...
Additional arguments, passed to pkg_deps$new().

Details

new_pkg_deps() creates a new object from the pkg_deps class. The advantage of new_pkg_deps() compared to using the pkg_deps constructor directly is that it avoids making pkgdepends a build time dependency.

The usual steps to query package dependencies are:

Create a pkg_deps object with new_pkg_deps().

Resolve all possible dependencies with pkg_deps$resolve().

Solve the dependencies, to obtain a subset of all possible dependencies that can be installed together, with pkg_deps$solve().

Call pkg_deps$get_solution() to list the result of the dependency solver.

Methods

Public methods

pkg_deps$new()

pkg_deps$get_refs()

pkg_deps$get_config()

pkg_deps$resolve()

pkg_deps$async_resolve()

pkg_deps$get_resolution()

pkg_deps$solve()

pkg_deps$get_solution()

pkg_deps$stop_for_solution_error()

pkg_deps$draw()

pkg_deps$format()

pkg_deps$print()

pkg_deps$clone()

Method new()

Create a new pkg_deps object. Consider using new_pkg_deps() instead of calling the constructor directly.

The returned object can be used to look up (recursive) dependencies of R packages from various sources. To perform the actual lookup, you'll need to call the resolve() method.

Usage

pkg_deps$new(refs, config = list(), remote_types = NULL)

Arguments

refs

Package names or references. See 'Package references' for the syntax.

config

Configuration options, a named list. See 'Configuration'.

remote_types

Custom remote ref types, this is for advanced use, and experimental currently.

Returns

A new pkg_deps object.

Examples

pd <- pkg_deps$new("r-lib/pkgdepends")
pd

Method get_refs()

The package refs that were used to create the pkg_deps object.

Usage

pkg_deps$get_refs()

Returns

A character vector of package refs that were used to create the pkg_deps object.

Examples

pd <- new_pkg_deps(c("pak", "jsonlite"))
pd$get_refs()

Method get_config()

Configuration options for the pkg_deps object. See 'Configuration' for details.

Usage

pkg_deps$get_config()

Returns

Named list. See 'Configuration' for the configuration options.

Examples

pd <- new_pkg_deps("pak")
pd$get_config()

Method resolve()

Resolve the dependencies of the specified package references. This usually means downloading metadata from CRAN and Bioconductor, unless already cached, and also from GitHub if GitHub refs were included, either directly or indirectly. See 'Dependency resolution' for details.

Usage

pkg_deps$resolve()

Returns

The pkg_deps object itself, invisibly.

Examples

pd <- new_pkg_deps("pak")
pd$resolve()
pd$get_resolution()

Method async_resolve()

The same as resolve(), but asynchronous. This method is for advanced use.

Usage

pkg_deps$async_resolve()

Returns

A deferred value.

Method get_resolution()

Query the result of the dependency resolution. This method can be called after resolve() has completed.

Usage

pkg_deps$get_resolution()

Returns

A pkg_resolution_result object, which is also a tibble. See 'Dependency resolution' for its columns.

Examples

pd <- new_pkg_deps("r-lib/pkgdepends")
pd$resolve()
pd$get_resolution()

Method solve()

Solve the package dependencies. Out of the resolved dependencies, it works out a set of packages, that can be installed together to create a functional installation. The set includes all directly specified packages, and all required (or suggested, depending on the configuration) packages as well. It includes every package at most once. See 'The dependency solver' for details.

solve() calls resolve() automatically, if it hasn't been called yet.

Usage

pkg_deps$solve()

Returns

The pkg_deps object itself, invisibly.

Examples

pd <- new_pkg_deps("r-lib/pkgdepends")
pd$resolve()
pd$solve()
pd$get_solution()

Method get_solution()

Returns the solution of the package dependencies.

Usage

pkg_deps$get_solution()

Returns

A pkg_solution_result object, which is a list. See pkg_solution_result for details.

Examples

pd <- new_pkg_deps("pkgload")
pd$resolve()
pd$solve()
pd$get_solution()

Method stop_for_solution_error()

Error if the dependency solver failed to find a consistent set of packages that can be installed together.

Usage

pkg_deps$stop_for_solution_error()

Examples

# This is an error, because the packages conflict:
pd <- new_pkg_deps(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)
pd$resolve()
pd$solve()
pd
# This fails:
# pd$stop_for_solution_error()

Method draw()

Draw a tree of package dependencies. It returns a tree object, see cli::tree(). Printing this object prints the dependency tree to the screen.

Usage

pkg_deps$draw()

Returns

A tree object from the cli package, see cli::tree().

Examples

pd <- new_pkg_deps("pkgload")
pd$solve()
pd$draw()

Method format()

Format a pkg_deps object, typically for printing.

Usage

pkg_deps$format(...)

Arguments

...

Not used currently.

Returns

A character vector, each element should be a line in the printout.

Method print()

Prints a pkg_deps object to the screen. The printout includes:

The package refs.

Whether the object has the resolved dependencies.

Whether the resolution had errors.

Whether the object has the solved dependencies.

Whether the solution had errors.

Advice on which methods to call next.

See the example below.

Usage

pkg_deps$print(...)

Arguments

...

not used currently.

Returns

The pkg_deps object itself, invisibly.

Examples

pd <- new_pkg_deps("r-lib/pkgdepends")
pd

pd$resolve()
pd

pd$solve()
pd

Method clone()

The objects of this class are cloneable with this method.

Usage

pkg_deps$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

pd <- pkg_deps$new("r-lib/pkgdepends")
pd

#> <pkg_dependencies>
#> + refs:
#>   - r-lib/pkgdepends
#> (use `$resolve()` to resolve dependencies)
#> (use `$solve()` to solve dependencies)

pd <- new_pkg_deps(c("pak", "jsonlite"))
pd$get_refs()

#> [1] "pak"      "jsonlite"

pd <- new_pkg_deps("pak")
pd$get_config()

#> + config:
#>   - library: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmpz4fPK7/file1da74a8880d5
#>   - cache_dir: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmpz4fPK7/file1da714d8a4fb
#>   - package_cache_dir: <default>
#>   - metadata_cache_dir: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmpz4fPK7/file1da71b37656e
#>   - platforms: macos, source
#>   - cran-mirror: https://cloud.r-project.org
#>   - dependencies: direct: Depends, Imports, LinkingTo, indirect: Depends, Imports, LinkingTo
#>   - r-versions: 4.0.2
#>   - build-vignettes: FALSE

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pd <- new_pkg_deps("pak")
pd$resolve()
pd$get_resolution()
}) # examplesIf

#> > pd <- new_pkg_deps("pak")
#> > pd$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [22.50 kB / 317.07 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.05 MB / 2.34 MB]   

#> ⠸ Updating metadata [0/12] | Downloading [2.07 MB / 2.34 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pd <- new_pkg_deps("r-lib/pkgdepends")
pd$resolve()
pd$get_resolution()
}) # examplesIf

#> > pd <- new_pkg_deps("r-lib/pkgdepends")
#> > pd$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [22.52 kB / 189.08 kB]

#> ⠹ Updating metadata [0/12] | Downloading [1.06 MB / 1.34 MB]   

#> ⠸ Updating metadata [0/12] | Downloading [2.12 MB / 2.34 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pd <- new_pkg_deps("r-lib/pkgdepends")
pd$resolve()
pd$solve()
pd$get_solution()
}) # examplesIf

#> > pd <- new_pkg_deps("r-lib/pkgdepends")
#> > pd$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [20 B / 207.10 kB]

#> ⠹ Updating metadata [0/12] | Downloading [1.04 MB / 1.34 MB]

#> ⠸ Updating metadata [0/12] | Downloading [2.11 MB / 2.34 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pd <- new_pkg_deps("pkgload")
pd$resolve()
pd$solve()
pd$get_solution()
}) # examplesIf

#> > pd <- new_pkg_deps("pkgload")
#> > pd$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [15.88 kB / 369.54 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.06 MB / 2.34 MB]   

#> ⠸ Updating metadata [0/12] | Downloading [2.09 MB / 2.34 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
# This is an error, because the packages conflict:
pd <- new_pkg_deps(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)
pd$resolve()
pd$solve()
pd
# This fails:
# pd$stop_for_solution_error()
}) # examplesIf

#> > pd <- new_pkg_deps(c("r-lib/pak", "cran::pak"), config = list(library = tempfile()))
#> > pd$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠹ Updating metadata [0/12] | Downloading [1.02 MB / 1.34 MB]

#> ⠸ Updating metadata [0/12] | Downloading [2.09 MB / 2.34 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pd <- new_pkg_deps("pkgload")
pd$solve()
pd$draw()
}) # examplesIf

#> > pd <- new_pkg_deps("pkgload")
#> > pd$solve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [31.75 kB / 221.58 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.09 MB / 2.34 MB]   

#> ⠸ Updating metadata [0/12] | Downloading [2.09 MB / 2.34 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pd <- new_pkg_deps("r-lib/pkgdepends")
pd

pd$resolve()
pd

pd$solve()
pd
}) # examplesIf

#> > pd <- new_pkg_deps("r-lib/pkgdepends")
#> > pd
#> <pkg_dependencies>
#> + refs:
#>   - r-lib/pkgdepends
#> (use `$resolve()` to resolve dependencies)
#> (use `$solve()` to solve dependencies)
#> > pd$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [22.50 kB / 337.04 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.07 MB / 2.34 MB]   

#> ⠸ Updating metadata [0/12] | Downloading [2.09 MB / 2.34 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).