R6 class for package download and installation.

Download and install R packages, with their dependencies, from various sources.

new_pkg_installation_proposal(refs, config = list(), ...)

Arguments

refs
Package names or references. See 'Package references' for the syntax.

config
Configuration options, a named list. See 'Configuration'. If it does not include library, then .libPaths()[1] is added as library.

...
Additional arguments, passed to pkg_installation_proposal$new().

Details

new_pkg_installation_proposal() creates a new object from the pkg_installation_proposal class. The advantage of new_pkg_installation_proposal() compared to using the pkg_installation_proposal constructor directly is that it avoids making pkgdepends a build time dependency.

Typical workflow to install a set of packages:

Create a pkg_installation_proposal object with new_pkg_installation_proposal().

Resolve all possible dependencies with pkg_installation_proposal$resolve().

Solve the package dependencies, to get an installation plan, with pkg_installation_proposal$solve().

Download all files with pkg_installation_proposal$download().

Install the downloaded files with pkg_installation_proposal$install().

Methods

Public methods

pkg_installation_proposal$new()

pkg_installation_proposal$get_refs()

pkg_installation_proposal$get_config()

pkg_installation_proposal$resolve()

pkg_installation_proposal$async_resolve()

pkg_installation_proposal$get_resolution()

pkg_installation_proposal$get_solve_policy()

pkg_installation_proposal$set_solve_policy()

pkg_installation_proposal$solve()

pkg_installation_proposal$get_solution()

pkg_installation_proposal$stop_for_solution_error()

pkg_installation_proposal$draw()

pkg_installation_proposal$download()

pkg_installation_proposal$async_download()

pkg_installation_proposal$get_downloads()

pkg_installation_proposal$stop_for_download_error()

pkg_installation_proposal$install()

pkg_installation_proposal$get_install_plan()

pkg_installation_proposal$format()

pkg_installation_proposal$print()

pkg_installation_proposal$clone()

Method new()

Create a new pkg_installation_proposal object. Consider using new_pkg_installation_proposal() instead of calling the constructor directly.

The returned object can be used to look up (recursive) dependencies of R packages from various sources, and then download and install the package files.

Usage

pkg_installation_proposal$new(
  refs,
  config = list(),
  policy = c("lazy", "upgrade"),
  remote_types = NULL
)

Arguments

refs

Package names or references. See 'Package references' for the syntax.

config

Configuration options, a named list. See 'Configuration'. It needs to include the package library to install to, in library.

policy

Solution policy. See 'The dependency solver'.

remote_types

Custom remote ref types, this is for advanced use, and experimental currently.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile()))
pdi

Method get_refs()

The package refs that were used to create the pkg_installation_proposal object.

Usage

pkg_installation_proposal$get_refs()

Returns

A character vector of package refs that were used to create the pkg_installation_proposal object.

Examples

pdi <- new_pkg_installation_proposal("r-lib/pkgdepends")
pdi$get_refs()

Method get_config()

Configuration options for the pkg_installation_proposal object. See 'Configuration' for details.

Usage

pkg_installation_proposal$get_config()

Returns

Named list. See 'Configuration' for the configuration options.

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi$get_config()

Method resolve()

Resolve the dependencies of the specified package references. This usually means downloading metadata from CRAN and Bioconductor, unless already cached, and also from GitHub if GitHub refs were included, either directly or indirectly. See 'Dependency resolution' for details.

Usage

pkg_installation_proposal$resolve()

Returns

The pkg_installation_proposal object, invisibly.

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

pdi$resolve()
pdi$get_resolution()

Method async_resolve()

The same as resolve(), but asynchronous. This method is for advanced use.

Usage

pkg_installation_proposal$async_resolve()

Returns

A deferred value.

Method get_resolution()

Query the result of the dependency resolution. This method can be called after resolve() has completed.

Usage

pkg_installation_proposal$get_resolution()

Returns

A pkg_resolution_result object, which is also a tibble. See 'Dependency resolution' for its columns.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$get_resolution()

Method get_solve_policy()

Returns the current policy of the dependency solver. See 'The dependency solver' for details.

Usage

pkg_installation_proposal$get_solve_policy()

Returns

A character vector of length one.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$get_solve_policy()
pdi$set_solve_policy("upgrade")
pdi$get_solve_policy()

Method set_solve_policy()

Set the current policy of the dependency solver. If the object already contains a solution and the new policy is different than the old policy, then the solution is deleted. See 'The dependency solver' for details.

Usage

pkg_installation_proposal$set_solve_policy(policy = c("lazy", "upgrade"))

Arguments

policy

Policy to set.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$get_solve_policy()
pdi$set_solve_policy("upgrade")
pdi$get_solve_policy()

Method solve()

Solve the package dependencies. Out of the resolved dependencies, it works out a set of packages, that can be installed together to create a functional installation. The set includes all directly specified packages, and all required (or suggested, depending on the configuration) packages as well. It includes every package at most once. See 'The dependency solver' for details.

Usage

pkg_installation_proposal$solve()

Returns

The pkg_installation_proposal object itself, invisibly.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$get_solution()

Method get_solution()

Returns the solution of the package dependencies.

Usage

pkg_installation_proposal$get_solution()

Returns

A pkg_solution_result object, which is a list. See pkg_solution_result for details.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$get_solution()

Method stop_for_solution_error()

Error if the dependency solver failed to find a consistent set of packages that can be installed together.

Usage

pkg_installation_proposal$stop_for_solution_error()

Examples

# This is an error, because the packages conflict:
pdi <- new_pkg_installation_proposal(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi
# This fails:
# pdi$stop_for_solution_error()

Method draw()

Draw a tree of package dependencies. It returns a tree object, see cli::tree(). Printing this object prints the dependency tree to the screen.

Usage

pkg_installation_proposal$draw()

Returns

A tree object from the cli package, see cli::tree().

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$draw()

Method download()

Download all packages that are part of the solution. It uses the package cache in the pkgcache package by default, to avoid downloads if possible.

Usage

pkg_installation_proposal$download()

Returns

The pkg_installation_proposal object itself, invisibly.

Examples

pdi <- new_pkg_installation_proposal(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$download()
pdi$get_downloads()

Method async_download()

The same as download(), but asynchronous. This method is for advanced use.

Usage

pkg_installation_proposal$async_download()

Returns

A deferred value.

Method get_downloads()

Returns the summary of the package downloads.

Usage

pkg_installation_proposal$get_downloads()

Returns

A pkg_download_result object, which is a list. See pkg_download_result for details.

Examples

pdi <- new_pkg_installation_proposal(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$download()
pdi$get_downloads()

Method stop_for_download_error()

Throw and error if the some of the downloads have failed for the most recent pkg_installation_proposal$download() call.

Usage

pkg_installation_proposal$stop_for_download_error()

Method install()

Install the downloaded packages. It calls install_package_plan().

Usage

pkg_installation_proposal$install()

Returns

The return value of install_package_plan().

Method get_install_plan()

Create an installation plan for the downloaded packages.

Usage

pkg_installation_proposal$get_install_plan()

Returns

An installation plan, see 'Installation plans' for the format.

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$download()
pdi$get_install_plan()

Method format()

Format a pkg_installation_proposal object, typically for printing.

Usage

pkg_installation_proposal$format(...)

Arguments

...

not used currently.

Returns

A character vector, each element should be a line in the printout.

Method print()

Prints a pkg_installation_proposal object to the screen.

The printout includes:

The package refs.

The policy of the dependency solver.

Whether the object has the solved dependencies.

Whether the solution had errors.

Whether the object has downloads.

Whether the downloads had errors.

Advice on which methods to call next.

See the example below.

Usage

pkg_installation_proposal$print(...)

Arguments

...

not used currently.

Returns

The pkg_installation_proposal object itself, invisibly.

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi

pdi$resolve()
pdi

pdi$solve()
pdi

pdi$download()
pdi

Method clone()

The objects of this class are cloneable with this method.

Usage

pkg_installation_proposal$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi

pdi$resolve()
pdi

pdi$solve()
pdi

pdi$download()
pdi
}) # examplesIf

#> > pdi <- new_pkg_installation_proposal("pak", config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f06204d093

#> > pdi
#> <pkg_installation_proposal>
#> + refs:
#>   - pak
#> + solution policy: lazy
#> (use `$solve()` to solve dependencies)
#> > pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [63.47 kB / 362.80 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.07 MB / 2.31 MB]   

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f03d763d1d

pdi

#> <pkg_installation_proposal>
#> + refs:
#>   - r-lib/pkgdepends
#> + solution policy: lazy
#> (use `$solve()` to solve dependencies)

pdi <- new_pkg_installation_proposal("r-lib/pkgdepends")
pdi$get_refs()

#> [1] "r-lib/pkgdepends"

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f0357187d0

pdi$get_config()

#> + config:
#>   - library: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f0357187d0
#>   - cache_dir: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f0528db5a7
#>   - package_cache_dir: <default>
#>   - metadata_cache_dir: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f055911148
#>   - platforms: macos, source
#>   - cran-mirror: https://cloud.r-project.org
#>   - dependencies: direct: Depends, Imports, LinkingTo, indirect: Depends, Imports, LinkingTo
#>   - r-versions: 4.0.1
#>   - build-vignettes: FALSE

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

pdi$resolve()
pdi$get_resolution()
}) # examplesIf

#> > pdi <- new_pkg_installation_proposal("pak", config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f02705b9db

#> > pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [47.60 kB / 342.83 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.06 MB / 2.31 MB]   

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$get_resolution()
}) # examplesIf

#> > pdi <- new_pkg_installation_proposal("r-lib/pkgdepends", config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f0876071d

#> > pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [31.74 kB / 194.86 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.07 MB / 2.31 MB]   

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f0d78a50c

pdi$get_solve_policy()

#> [1] "lazy"

pdi$set_solve_policy("upgrade")
pdi$get_solve_policy()

#> [1] "upgrade"

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f063b8b745

pdi$get_solve_policy()

#> [1] "lazy"

pdi$set_solve_policy("upgrade")
pdi$get_solve_policy()

#> [1] "upgrade"

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$get_solution()
}) # examplesIf

#> > pdi <- new_pkg_installation_proposal("r-lib/pkgdepends", config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f04c254114

#> > pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [0 B / 162.34 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.01 MB / 2.31 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$get_solution()
}) # examplesIf

#> > pdi <- new_pkg_installation_proposal("r-lib/pkgdepends", config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f0f065fd2

#> > pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [31.73 kB / 52.47 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.03 MB / 2.31 MB]  

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
# This is an error, because the packages conflict:
pdi <- new_pkg_installation_proposal(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi
# This fails:
# pdi$stop_for_solution_error()
}) # examplesIf

#> > pdi <- new_pkg_installation_proposal(c("r-lib/pak", "cran::pak"), config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f04224f786

#> > pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [22.50 kB / 188.95 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.07 MB / 2.31 MB]   

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$draw()
}) # examplesIf

#> > pdi <- new_pkg_installation_proposal("pak", config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f0647c471d

#> > pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [20 B / 20 B]

#> ⠹ Updating metadata [0/12] | Downloading [2.01 MB / 2.31 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pdi <- new_pkg_installation_proposal(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$download()
pdi$get_downloads()
}) # examplesIf

#> > pdi <- new_pkg_installation_proposal(c("r-lib/pak", "cran::pak"), config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f0528cc239

#> > pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠹ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠸ Updating metadata [0/12] | Downloading [2.07 MB / 2.31 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pdi <- new_pkg_installation_proposal(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$download()
pdi$get_downloads()
}) # examplesIf

#> > pdi <- new_pkg_installation_proposal(c("r-lib/pak", "cran::pak"), config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f05c2f2fa4

#> > pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [6.66 kB / 6.66 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.00 MB / 2.15 MB]

#> ⠸ Updating metadata [0/12] | Downloading [2.05 MB / 2.15 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$download()
pdi$get_install_plan()
}) # examplesIf

#> > pdi <- new_pkg_installation_proposal("pak", config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f02c43ef3e

#> > pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [15.88 kB / 59.13 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.02 MB / 2.31 MB]  

#> ⠸ Updating metadata [0/12] | Downloading [2.07 MB / 2.31 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).

if (pkgdepends:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi

pdi$resolve()
pdi

pdi$solve()
pdi

pdi$download()
pdi
}) # examplesIf

#> > pdi <- new_pkg_installation_proposal("pak", config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpXkVCrR/file7f06d148c4e

#> > pdi
#> <pkg_installation_proposal>
#> + refs:
#>   - pak
#> + solution policy: lazy
#> (use `$solve()` to solve dependencies)
#> > pdi$resolve()

#> Checking 12 metadata files

#> ⠋ Updating metadata [0/12] | Downloading [0 B / 0 B]

#> ⠙ Updating metadata [0/12] | Downloading [0 B / 32.50 kB]

#> ⠹ Updating metadata [0/12] | Downloading [2.01 MB / 2.31 MB]

#> ✖ Metadata update failed                                    

#> 

#> Error in stop(http_error(resp)): Not Found (HTTP 404).