R6 class for package download and installation.

Download and install R packages, with their dependencies, from various sources.

new_pkg_installation_proposal(refs, config = list(), ...)

Arguments

refs
Package names or references. See 'Package references' for the syntax.

config
Configuration options, a named list. See 'Configuration'. If it does not include library, then .libPaths()[1] is added as library.

...
Additional arguments, passed to pkg_installation_proposal$new().

Details

new_pkg_installation_proposal() creates a new object from the pkg_installation_proposal class. The advantage of new_pkg_installation_proposal() compared to using the pkg_installation_proposal constructor directly is that it avoids making pkgdepends a build time dependency.

Typical workflow to install a set of packages:

Create a pkg_installation_proposal object with new_pkg_installation_proposal().

Resolve all possible dependencies with pkg_installation_proposal$resolve().

Solve the package dependencies, to get an installation plan, with pkg_installation_proposal$solve().

Download all files with pkg_installation_proposal$download().

Install the downloaded files with pkg_installation_proposal$install().

Methods

Public methods

pkg_installation_proposal$new()

pkg_installation_proposal$get_refs()

pkg_installation_proposal$get_config()

pkg_installation_proposal$resolve()

pkg_installation_proposal$async_resolve()

pkg_installation_proposal$get_resolution()

pkg_installation_proposal$get_solve_policy()

pkg_installation_proposal$set_solve_policy()

pkg_installation_proposal$solve()

pkg_installation_proposal$get_solution()

pkg_installation_proposal$show_solution()

pkg_installation_proposal$stop_for_solution_error()

pkg_installation_proposal$draw()

pkg_installation_proposal$download()

pkg_installation_proposal$async_download()

pkg_installation_proposal$get_downloads()

pkg_installation_proposal$stop_for_download_error()

pkg_installation_proposal$install()

pkg_installation_proposal$get_install_plan()

pkg_installation_proposal$format()

pkg_installation_proposal$print()

pkg_installation_proposal$clone()

Method new()

Create a new pkg_installation_proposal object. Consider using new_pkg_installation_proposal() instead of calling the constructor directly.

The returned object can be used to look up (recursive) dependencies of R packages from various sources, and then download and install the package files.

Usage

pkg_installation_proposal$new(
  refs,
  config = list(),
  policy = c("lazy", "upgrade"),
  remote_types = NULL
)

Arguments

refs

Package names or references. See 'Package references' for the syntax.

config

Configuration options, a named list. See 'Configuration'. It needs to include the package library to install to, in library.

policy

Solution policy. See 'The dependency solver'.

remote_types

Custom remote ref types, this is for advanced use, and experimental currently.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile()))
pdi

Method get_refs()

The package refs that were used to create the pkg_installation_proposal object.

Usage

pkg_installation_proposal$get_refs()

Returns

A character vector of package refs that were used to create the pkg_installation_proposal object.

Examples

pdi <- new_pkg_installation_proposal("r-lib/pkgdepends")
pdi$get_refs()

Method get_config()

Configuration options for the pkg_installation_proposal object. See 'Configuration' for details.

Usage

pkg_installation_proposal$get_config()

Returns

Named list. See 'Configuration' for the configuration options.

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi$get_config()

Method resolve()

Resolve the dependencies of the specified package references. This usually means downloading metadata from CRAN and Bioconductor, unless already cached, and also from GitHub if GitHub refs were included, either directly or indirectly. See 'Dependency resolution' for details.

Usage

pkg_installation_proposal$resolve()

Returns

The pkg_installation_proposal object, invisibly.

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

pdi$resolve()
pdi$get_resolution()

Method async_resolve()

The same as resolve(), but asynchronous. This method is for advanced use.

Usage

pkg_installation_proposal$async_resolve()

Returns

A deferred value.

Method get_resolution()

Query the result of the dependency resolution. This method can be called after resolve() has completed.

Usage

pkg_installation_proposal$get_resolution()

Returns

A pkg_resolution_result object, which is also a tibble. See 'Dependency resolution' for its columns.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$get_resolution()

Method get_solve_policy()

Returns the current policy of the dependency solver. See 'The dependency solver' for details.

Usage

pkg_installation_proposal$get_solve_policy()

Returns

A character vector of length one.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$get_solve_policy()
pdi$set_solve_policy("upgrade")
pdi$get_solve_policy()

Method set_solve_policy()

Set the current policy of the dependency solver. If the object already contains a solution and the new policy is different than the old policy, then the solution is deleted. See 'The dependency solver' for details.

Usage

pkg_installation_proposal$set_solve_policy(policy = c("lazy", "upgrade"))

Arguments

policy

Policy to set.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$get_solve_policy()
pdi$set_solve_policy("upgrade")
pdi$get_solve_policy()

Method solve()

Solve the package dependencies. Out of the resolved dependencies, it works out a set of packages, that can be installed together to create a functional installation. The set includes all directly specified packages, and all required (or suggested, depending on the configuration) packages as well. It includes every package at most once. See 'The dependency solver' for details.

Usage

pkg_installation_proposal$solve()

Returns

The pkg_installation_proposal object itself, invisibly.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$get_solution()

Method get_solution()

Returns the solution of the package dependencies.

Usage

pkg_installation_proposal$get_solution()

Returns

A pkg_solution_result object, which is a list. See pkg_solution_result for details.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$get_solution()

Method show_solution()

Show the solution on the screen.

Usage

pkg_installation_proposal$show_solution(key = FALSE)

Arguments

key

Whether to show the key to the package list annotation.

Returns

A pkg_solution_result object, which is a list. See pkg_solution_result for details.

Examples

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$get_solution()
pdi$show_solution()

Method stop_for_solution_error()

Error if the dependency solver failed to find a consistent set of packages that can be installed together.

Usage

pkg_installation_proposal$stop_for_solution_error()

Examples

# This is an error, because the packages conflict:
pdi <- new_pkg_installation_proposal(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi
# This fails:
# pdi$stop_for_solution_error()

Method draw()

Draw a tree of package dependencies. It returns a tree object, see cli::tree(). Printing this object prints the dependency tree to the screen.

Usage

pkg_installation_proposal$draw()

Returns

A tree object from the cli package, see cli::tree().

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$draw()

Method download()

Download all packages that are part of the solution. It uses the package cache in the pkgcache package by default, to avoid downloads if possible.

Usage

pkg_installation_proposal$download()

Returns

The pkg_installation_proposal object itself, invisibly.

Examples

pdi <- new_pkg_installation_proposal(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$download()
pdi$get_downloads()

Method async_download()

The same as download(), but asynchronous. This method is for advanced use.

Usage

pkg_installation_proposal$async_download()

Returns

A deferred value.

Method get_downloads()

Returns the summary of the package downloads.

Usage

pkg_installation_proposal$get_downloads()

Returns

A pkg_download_result object, which is a list. See pkg_download_result for details.

Examples

pdi <- new_pkg_installation_proposal(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$download()
pdi$get_downloads()

Method stop_for_download_error()

Throw and error if the some of the downloads have failed for the most recent pkg_installation_proposal$download() call.

Usage

pkg_installation_proposal$stop_for_download_error()

Method install()

Install the downloaded packages. It calls install_package_plan().

Usage

pkg_installation_proposal$install()

Returns

The return value of install_package_plan().

Method get_install_plan()

Create an installation plan for the downloaded packages.

Usage

pkg_installation_proposal$get_install_plan()

Returns

An installation plan, see 'Installation plans' for the format.

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi$resolve()
pdi$solve()
pdi$download()
pdi$get_install_plan()

Method format()

Format a pkg_installation_proposal object, typically for printing.

Usage

pkg_installation_proposal$format(...)

Arguments

...

not used currently.

Returns

A character vector, each element should be a line in the printout.

Method print()

Prints a pkg_installation_proposal object to the screen.

The printout includes:

The package refs.

The policy of the dependency solver.

Whether the object has the solved dependencies.

Whether the solution had errors.

Whether the object has downloads.

Whether the downloads had errors.

Advice on which methods to call next.

See the example below.

Usage

pkg_installation_proposal$print(...)

Arguments

...

not used currently.

Returns

The pkg_installation_proposal object itself, invisibly.

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)
pdi

pdi$resolve()
pdi

pdi$solve()
pdi

pdi$download()
pdi

Method clone()

The objects of this class are cloneable with this method.

Usage

pkg_installation_proposal$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file759c747725

pdi

#> <pkg_installation_proposal>
#> + refs:
#>   - pak
#> + solution policy: lazy
#> (use `$solve()` to solve dependencies)

pdi$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving standard (CRAN/BioC) packages

#>                                                                           

pdi

#> <pkg_installation_proposal>
#> + refs:
#>   - pak
#> + solution policy: lazy
#> (use `$solve()` to solve dependencies)

pdi$solve()
pdi

#> <pkg_installation_proposal>
#> + refs:
#>   - pak
#> + solution policy: lazy
#> + has solution
#> (use `$download()` to download packages)
#> (use `$show_solution()` to see the packages to install
#> (use `$get_solution()` to see the full solution results)
#> (use `$draw()` to draw the dependency tree)

pdi$download()

#> ℹ Getting 25 pkgs (7.84 MB), 18 (20.92 MB) cached

#> 

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of assertthat 0.2.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of base64enc 0.1-3 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of crayon 1.3.4 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of curl 4.3 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of desc 1.2.0 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of ellipsis 0.3.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of fansi 0.4.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of filelock 1.0.2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of hms 0.5.3 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of lifecycle 0.2.0 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of lpSolve 5.6.15 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of magrittr 1.5 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of pak 0.1.2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of pkgconfig 2.0.3 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of prettycode 1.1.0 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of prettyunits 1.1.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of progress 1.2.2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of rappdirs 0.3.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of rematch2 2.1.2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of rprojroot 1.3-2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of selectr 0.4-2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of stringr 1.4.0 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of utf8 1.1.4 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of uuid 0.1-4 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of xml2 1.3.2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>  0.0 kB/s ⸨█████████████⸩ 100% | 25/25 pkg | ETA ??s  | Connecting...

#>  0.0 kB/s ⸨█████████████⸩ 100% | 25/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ No downloads needed, all packages are cached

pdi

#> <pkg_installation_proposal>
#> + refs:
#>   - pak
#> + solution policy: lazy
#> + has solution
#> + has downloads
#> (use `$show_solution()` to see the packages to install
#> (use `$get_solution()` to see the full solution results)
#> (use `$draw()` to draw the dependency tree)
#> (use `$get_downloads()` to get download data)
#> (use `$get_install_plan()` to get the installation plan)
#> (use `$install()` to install the packages)

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile()))

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file75928489ed4

pdi

#> <pkg_installation_proposal>
#> + refs:
#>   - r-lib/pkgdepends
#> + solution policy: lazy
#> (use `$solve()` to solve dependencies)

pdi <- new_pkg_installation_proposal("r-lib/pkgdepends")
pdi$get_refs()

#> [1] "r-lib/pkgdepends"

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file75976c3d0ef

pdi$get_config()

#> + config:
#>   - library: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file75976c3d0ef
#>   - cache_dir: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file75910f2f7a0
#>   - package_cache_dir: <default>
#>   - metadata_cache_dir: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file7593f5f3211
#>   - platforms: macos, source
#>   - cran-mirror: https://cloud.r-project.org
#>   - dependencies: direct: Depends, Imports, LinkingTo, indirect: Depends, Imports, LinkingTo
#>   - r-versions: 4.0.3
#>   - build-vignettes: FALSE

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file7592d01a98

pdi$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving standard (CRAN/BioC) packages

#>                                                                           

pdi$get_resolution()

#> # A tibble: 86 x 27
#>    ref   type  direct directpkg status package version license needscompilation
#>    <chr> <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 R6    stan… FALSE  FALSE     OK     R6      2.5.0   MIT + … FALSE           
#>  2 asse… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  3 back… stan… FALSE  FALSE     OK     backpo… 1.2.0   GPL-2 … FALSE           
#>  4 base… stan… FALSE  FALSE     OK     base64… 0.1-3   GPL-2 … FALSE           
#>  5 callr stan… FALSE  FALSE     OK     callr   3.5.1   MIT + … FALSE           
#>  6 cli   stan… FALSE  FALSE     OK     cli     2.1.0   MIT + … FALSE           
#>  7 clia… stan… FALSE  FALSE     OK     cliapp  0.1.1   MIT + … FALSE           
#>  8 cray… stan… FALSE  FALSE     OK     crayon  1.3.4   MIT + … FALSE           
#>  9 curl  stan… FALSE  FALSE     OK     curl    4.3     MIT + … FALSE           
#> 10 desc  stan… FALSE  FALSE     OK     desc    1.2.0   MIT + … FALSE           
#> # … with 76 more rows, and 18 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repodir <chr>, target <chr>, deps <list>, mirror <chr>, sources <list>,
#> #   remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, cache_status <chr>

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file759e09a389

pdi$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving r-lib/pkgdepends

#>  Found  1  deps for  0/1  pkgs [⠙] Resolving r-lib/pkgdepends

#>  Found  1  deps for  0/1  pkgs [⠹] Resolving r-lib/pkgdepends

#>  Found  21  deps for  1/1  pkgs [⠸] Resolving standard (CRAN/BioC) packages

#>                                                                            

pdi$get_resolution()

#> # A tibble: 69 x 27
#>    ref   type  direct directpkg status package version license needscompilation
#>    <chr> <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 r-li… gith… TRUE   TRUE      OK     pkgdep… 0.0.0.… MIT + … TRUE            
#>  2 R6    stan… FALSE  FALSE     OK     R6      2.5.0   MIT + … FALSE           
#>  3 asse… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  4 back… stan… FALSE  FALSE     OK     backpo… 1.2.0   GPL-2 … FALSE           
#>  5 callr stan… FALSE  FALSE     OK     callr   3.5.1   MIT + … FALSE           
#>  6 cli   stan… FALSE  FALSE     OK     cli     2.1.0   MIT + … FALSE           
#>  7 cray… stan… FALSE  FALSE     OK     crayon  1.3.4   MIT + … FALSE           
#>  8 curl  stan… FALSE  FALSE     OK     curl    4.3     MIT + … FALSE           
#>  9 desc  stan… FALSE  FALSE     OK     desc    1.2.0   MIT + … FALSE           
#> 10 dige… stan… FALSE  FALSE     OK     digest  0.6.27  GPL (>… FALSE           
#> # … with 59 more rows, and 18 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repodir <chr>, target <glue>, deps <list>, mirror <chr>, sources <list>,
#> #   remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, cache_status <chr>

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file759ebb9209

pdi$get_solve_policy()

#> [1] "lazy"

pdi$set_solve_policy("upgrade")
pdi$get_solve_policy()

#> [1] "upgrade"

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file7597036f613

pdi$get_solve_policy()

#> [1] "lazy"

pdi$set_solve_policy("upgrade")
pdi$get_solve_policy()

#> [1] "upgrade"

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file759273c534b

pdi$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving r-lib/pkgdepends

#>  Found  21  deps for  1/1  pkgs [⠙] Resolving standard (CRAN/BioC) packages

#>                                                                            

pdi$solve()
pdi$get_solution()

#> <pkg_solution>
#> + result: OK
#> + refs:
#>   - r-lib/pkgdepends
#> + constraints (150):
#>   - select pkgdepends exactly once
#>   - select R6 at most once
#>   - select assertthat at most once
#>   - select backports at most once
#>   - select callr at most once
#>   - select cli at most once
#>   - select crayon at most once
#>   - select curl at most once
#>   - select desc at most once
#>   - select digest at most once
#>   ...
#> + solution:
#>   - assertthat
#>   - backports
#>   - callr
#>   - cli
#>   - crayon
#>   - curl
#>   - desc
#>   - digest
#>   - ellipsis
#>   - fansi
#>   - filelock
#>   - glue
#>   - jsonlite
#>   - lifecycle
#>   - lpSolve
#>   - magrittr
#>   - pillar
#>   - pkgbuild
#>   - pkgcache
#>   - pkgconfig
#>   - prettyunits
#>   - processx
#>   - ps
#>   - r-lib/pkgdepends
#>   - R6
#>   - rappdirs
#>   - rematch2
#>   - rlang
#>   - rprojroot
#>   - tibble
#>   - utf8
#>   - uuid
#>   - vctrs
#>   - withr
#>   - zip

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file7593d509147

pdi$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving r-lib/pkgdepends

#>  Found  21  deps for  1/1  pkgs [⠙] Resolving standard (CRAN/BioC) packages

#>  Found  21  deps for  1/1  pkgs [⠹] Resolving standard (CRAN/BioC) packages

#>                                                                            

pdi$solve()
pdi$get_solution()

#> <pkg_solution>
#> + result: OK
#> + refs:
#>   - r-lib/pkgdepends
#> + constraints (150):
#>   - select pkgdepends exactly once
#>   - select R6 at most once
#>   - select assertthat at most once
#>   - select backports at most once
#>   - select callr at most once
#>   - select cli at most once
#>   - select crayon at most once
#>   - select curl at most once
#>   - select desc at most once
#>   - select digest at most once
#>   ...
#> + solution:
#>   - assertthat
#>   - backports
#>   - callr
#>   - cli
#>   - crayon
#>   - curl
#>   - desc
#>   - digest
#>   - ellipsis
#>   - fansi
#>   - filelock
#>   - glue
#>   - jsonlite
#>   - lifecycle
#>   - lpSolve
#>   - magrittr
#>   - pillar
#>   - pkgbuild
#>   - pkgcache
#>   - pkgconfig
#>   - prettyunits
#>   - processx
#>   - ps
#>   - r-lib/pkgdepends
#>   - R6
#>   - rappdirs
#>   - rematch2
#>   - rlang
#>   - rprojroot
#>   - tibble
#>   - utf8
#>   - uuid
#>   - vctrs
#>   - withr
#>   - zip

pdi <- new_pkg_installation_proposal(
  "r-lib/pkgdepends",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file7597399994d

pdi$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving r-lib/pkgdepends

#>  Found  21  deps for  1/1  pkgs [⠙] Resolving standard (CRAN/BioC) packages

#>                                                                            

pdi$solve()
pdi$get_solution()

#> <pkg_solution>
#> + result: OK
#> + refs:
#>   - r-lib/pkgdepends
#> + constraints (150):
#>   - select pkgdepends exactly once
#>   - select R6 at most once
#>   - select assertthat at most once
#>   - select backports at most once
#>   - select callr at most once
#>   - select cli at most once
#>   - select crayon at most once
#>   - select curl at most once
#>   - select desc at most once
#>   - select digest at most once
#>   ...
#> + solution:
#>   - assertthat
#>   - backports
#>   - callr
#>   - cli
#>   - crayon
#>   - curl
#>   - desc
#>   - digest
#>   - ellipsis
#>   - fansi
#>   - filelock
#>   - glue
#>   - jsonlite
#>   - lifecycle
#>   - lpSolve
#>   - magrittr
#>   - pillar
#>   - pkgbuild
#>   - pkgcache
#>   - pkgconfig
#>   - prettyunits
#>   - processx
#>   - ps
#>   - r-lib/pkgdepends
#>   - R6
#>   - rappdirs
#>   - rematch2
#>   - rlang
#>   - rprojroot
#>   - tibble
#>   - utf8
#>   - uuid
#>   - vctrs
#>   - withr
#>   - zip

pdi$show_solution()

#> + assertthat    0.2.1       ⬇ (52.47 kB)
#> + backports     1.2.0      
#> + callr         3.5.1      
#> + cli           2.1.0      
#> + crayon        1.3.4       ⬇ (748.25 kB)
#> + curl          4.3         ⬇ (741.06 kB)
#> + desc          1.2.0       ⬇ (279.60 kB)
#> + digest        0.6.27     
#> + ellipsis      0.3.1       ⬇ (33.48 kB)
#> + fansi         0.4.1       ⬇ (210.40 kB)
#> + filelock      1.0.2       ⬇ (26.67 kB)
#> + glue          1.4.2      
#> + jsonlite      1.7.1      
#> + lifecycle     0.2.0       ⬇ (91.64 kB)
#> + lpSolve       5.6.15      ⬇ (1.07 MB)
#> + magrittr      1.5         ⬇ (152.51 kB)
#> + pillar        1.4.6      
#> + pkgbuild      1.1.0      
#> + pkgcache      1.1.1      
#> + pkgconfig     2.0.3       ⬇ (17.63 kB)
#> + pkgdepends    0.0.0.9007 👷🏿🔧 ⬇ (GitHub: 2a804d6)
#> + prettyunits   1.1.1       ⬇ (34.79 kB)
#> + processx      3.4.4      
#> + ps            1.4.0      
#> + R6            2.5.0      
#> + rappdirs      0.3.1       ⬇ (145.56 kB)
#> + rematch2      2.1.2       ⬇ (44.87 kB)
#> + rlang         0.4.8      
#> + rprojroot     1.3-2       ⬇ (78.79 kB)
#> + tibble        3.0.4      
#> + utf8          1.1.4       ⬇ (195.28 kB)
#> + uuid          0.1-4       ⬇ (27.75 kB)
#> + vctrs         0.3.4      
#> + withr         2.3.0      
#> + zip           2.1.1      

# This is an error, because the packages conflict:
pdi <- new_pkg_installation_proposal(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file7595d5368e5

pdi$resolve()

#> 

#> 

#> 

#>  Found  2  deps for  0/2  pkgs [⠋] Resolving r-lib/pak

#>  Found  2  deps for  1/2  pkgs [⠙] Resolving r-lib/pak

#>                                                       

pdi$solve()
pdi

#> <pkg_installation_proposal>
#> + refs:
#>   - r-lib/pak
#>   - cran::pak
#> + solution policy: lazy
#> + has solution
#> x has solution errors
#> (use `$show_solution()` to see the packages to install
#> (use `$get_solution()` to see the full solution results)

# This fails:
# pdi$stop_for_solution_error()
pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file7594914ffc4

pdi$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving standard (CRAN/BioC) packages

#>                                                                           

pdi$solve()
pdi$draw()

#> pak 0.1.2 ✨ ⬇ (402.81 kB)
#> ├─assertthat 0.2.1 ✨ ⬇ (52.47 kB)
#> ├─base64enc 0.1-3 ✨ ⬇ (31.63 kB)
#> ├─callr 3.5.1 ✨
#> │ ├─processx 3.4.4 ✨
#> │ │ ├─ps 1.4.0 ✨
#> │ │ └─R6 2.5.0 ✨
#> │ └─R6
#> ├─cli 2.1.0 ✨
#> │ ├─assertthat
#> │ ├─crayon 1.3.4 ✨ ⬇ (748.25 kB)
#> │ ├─glue 1.4.2 ✨
#> │ └─fansi 0.4.1 ✨ ⬇ (210.40 kB)
#> ├─cliapp 0.1.1 ✨
#> │ ├─cli
#> │ ├─crayon
#> │ ├─fansi
#> │ ├─glue
#> │ ├─prettycode 1.1.0 ✨ ⬇ (243.25 kB)
#> │ │ └─crayon
#> │ ├─progress 1.2.2 ✨ ⬇ (81.79 kB)
#> │ │ ├─hms 0.5.3 ✨ ⬇ (103.27 kB)
#> │ │ │ ├─pkgconfig 2.0.3 ✨ ⬇ (17.63 kB)
#> │ │ │ ├─rlang 0.4.8 ✨
#> │ │ │ └─vctrs 0.3.4 ✨
#> │ │ │   ├─ellipsis 0.3.1 ✨ ⬇ (33.48 kB)
#> │ │ │   │ └─rlang
#> │ │ │   ├─digest 0.6.27 ✨
#> │ │ │   ├─glue
#> │ │ │   └─rlang
#> │ │ ├─prettyunits 1.1.1 ✨ ⬇ (34.79 kB)
#> │ │ ├─R6
#> │ │ └─crayon
#> │ ├─R6
#> │ ├─selectr 0.4-2 ✨ ⬇ (484.31 kB)
#> │ │ ├─stringr 1.4.0 ✨ ⬇ (213.23 kB)
#> │ │ │ ├─glue
#> │ │ │ ├─magrittr 1.5 ✨ ⬇ (152.51 kB)
#> │ │ │ └─stringi 1.5.3 ✨
#> │ │ └─R6
#> │ ├─withr 2.3.0 ✨
#> │ └─xml2 1.3.2 ✨ ⬇ (2.33 MB)
#> ├─crayon
#> ├─curl 4.3 ✨ ⬇ (741.06 kB)
#> ├─desc 1.2.0 ✨ ⬇ (279.60 kB)
#> │ ├─assertthat
#> │ ├─R6
#> │ ├─crayon
#> │ └─rprojroot 1.3-2 ✨ ⬇ (78.79 kB)
#> │   └─backports 1.2.0 ✨
#> ├─filelock 1.0.2 ✨ ⬇ (26.67 kB)
#> ├─glue
#> ├─jsonlite 1.7.1 ✨
#> ├─lpSolve 5.6.15 ✨ ⬇ (1.07 MB)
#> ├─pkgbuild 1.1.0 ✨
#> │ ├─callr
#> │ ├─cli
#> │ ├─crayon
#> │ ├─desc
#> │ ├─prettyunits
#> │ ├─R6
#> │ ├─rprojroot
#> │ └─withr
#> ├─pkgcache 1.1.1 ✨
#> │ ├─assertthat
#> │ ├─callr
#> │ ├─cli
#> │ ├─curl
#> │ ├─digest
#> │ ├─filelock
#> │ ├─glue
#> │ ├─prettyunits
#> │ ├─R6
#> │ ├─processx
#> │ ├─rappdirs 0.3.1 ✨ ⬇ (145.56 kB)
#> │ ├─rlang
#> │ ├─tibble 3.0.4 ✨
#> │ │ ├─cli
#> │ │ ├─crayon
#> │ │ ├─ellipsis
#> │ │ ├─fansi
#> │ │ ├─lifecycle 0.2.0 ✨ ⬇ (91.64 kB)
#> │ │ │ ├─glue
#> │ │ │ └─rlang
#> │ │ ├─magrittr
#> │ │ ├─pillar 1.4.6 ✨
#> │ │ │ ├─cli
#> │ │ │ ├─crayon
#> │ │ │ ├─ellipsis
#> │ │ │ ├─fansi
#> │ │ │ ├─lifecycle
#> │ │ │ ├─rlang
#> │ │ │ ├─utf8 1.1.4 ✨ ⬇ (195.28 kB)
#> │ │ │ └─vctrs
#> │ │ ├─pkgconfig
#> │ │ ├─rlang
#> │ │ └─vctrs
#> │ └─uuid 0.1-4 ✨ ⬇ (27.75 kB)
#> ├─prettyunits
#> ├─processx
#> ├─ps
#> ├─R6
#> ├─rematch2 2.1.2 ✨ ⬇ (44.87 kB)
#> │ └─tibble
#> ├─rprojroot
#> └─tibble
#> 
#> Key:  ✨ new |  ⬇ download

pdi <- new_pkg_installation_proposal(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file759547a1011

pdi$resolve()

#> 

#> 

#> 

#>  Found  2  deps for  0/2  pkgs [⠋] Resolving r-lib/pak

#>  Found  2  deps for  1/2  pkgs [⠙] Resolving r-lib/pak

#>                                                       

pdi$solve()
pdi$download()

#> ℹ No downloads are needed

#> 

#> 

pdi$get_downloads()

#> # A tibble: 0 x 35
#> # … with 35 variables: ref <chr>, type <chr>, direct <lgl>, directpkg <lgl>,
#> #   status <chr>, package <chr>, version <chr>, license <chr>,
#> #   needscompilation <lgl>, priority <chr>, md5sum <chr>, sha256 <chr>,
#> #   filesize <int>, built <chr>, platform <chr>, rversion <chr>, repodir <chr>,
#> #   target <chr>, deps <list>, mirror <chr>, sources <list>, remote <list>,
#> #   error <list>, metadata <list>, extra <list>, dep_types <list>,
#> #   cache_status <lgl>, lib_status <chr>, old_version <chr>, new_version <chr>,
#> #   fulltarget <chr>, fulltarget_tree <chr>, download_status <chr>,
#> #   download_error <list>, file_size <dbl>

pdi <- new_pkg_installation_proposal(
  c("r-lib/pak", "cran::pak"),
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file7592489fa46

pdi$resolve()

#> 

#> 

#> 

#>  Found  2  deps for  0/2  pkgs [⠋] Resolving r-lib/pak

#>  Found  2  deps for  1/2  pkgs [⠙] Resolving r-lib/pak

#>                                                       

pdi$solve()
pdi$download()

#> ℹ No downloads are needed

#> 

#> 

pdi$get_downloads()

#> # A tibble: 0 x 35
#> # … with 35 variables: ref <chr>, type <chr>, direct <lgl>, directpkg <lgl>,
#> #   status <chr>, package <chr>, version <chr>, license <chr>,
#> #   needscompilation <lgl>, priority <chr>, md5sum <chr>, sha256 <chr>,
#> #   filesize <int>, built <chr>, platform <chr>, rversion <chr>, repodir <chr>,
#> #   target <chr>, deps <list>, mirror <chr>, sources <list>, remote <list>,
#> #   error <list>, metadata <list>, extra <list>, dep_types <list>,
#> #   cache_status <lgl>, lib_status <chr>, old_version <chr>, new_version <chr>,
#> #   fulltarget <chr>, fulltarget_tree <chr>, download_status <chr>,
#> #   download_error <list>, file_size <dbl>

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file759104f069f

pdi$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving standard (CRAN/BioC) packages

#>                                                                           

pdi$solve()
pdi$download()

#> ℹ Getting 25 pkgs (7.84 MB), 18 (20.92 MB) cached

#> 

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of assertthat 0.2.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of base64enc 0.1-3 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of crayon 1.3.4 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of curl 4.3 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of desc 1.2.0 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of ellipsis 0.3.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of fansi 0.4.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of filelock 1.0.2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of hms 0.5.3 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of lifecycle 0.2.0 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of lpSolve 5.6.15 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of magrittr 1.5 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of pak 0.1.2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of pkgconfig 2.0.3 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of prettycode 1.1.0 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of prettyunits 1.1.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of progress 1.2.2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of rappdirs 0.3.1 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of rematch2 2.1.2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of rprojroot 1.3-2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of selectr 0.4-2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of stringr 1.4.0 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of utf8 1.1.4 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of uuid 0.1-4 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of xml2 1.3.2 (macos) is the latest build

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>  0.0 kB/s ⸨█████████████⸩ 100% | 25/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ No downloads needed, all packages are cached

pdi$get_install_plan()

#> # A tibble: 43 x 42
#>    ref   type  direct directpkg status package version license needscompilation
#>    <chr> <chr> <lgl>  <lgl>     <chr>  <chr>   <chr>   <chr>   <lgl>           
#>  1 R6    stan… FALSE  FALSE     OK     R6      2.5.0   MIT + … FALSE           
#>  2 asse… stan… FALSE  FALSE     OK     assert… 0.2.1   GPL-3   FALSE           
#>  3 back… stan… FALSE  FALSE     OK     backpo… 1.2.0   GPL-2 … FALSE           
#>  4 base… stan… FALSE  FALSE     OK     base64… 0.1-3   GPL-2 … FALSE           
#>  5 callr stan… FALSE  FALSE     OK     callr   3.5.1   MIT + … FALSE           
#>  6 cli   stan… FALSE  FALSE     OK     cli     2.1.0   MIT + … FALSE           
#>  7 clia… stan… FALSE  FALSE     OK     cliapp  0.1.1   MIT + … FALSE           
#>  8 cray… stan… FALSE  FALSE     OK     crayon  1.3.4   MIT + … FALSE           
#>  9 curl  stan… FALSE  FALSE     OK     curl    4.3     MIT + … FALSE           
#> 10 desc  stan… FALSE  FALSE     OK     desc    1.2.0   MIT + … FALSE           
#> # … with 33 more rows, and 33 more variables: priority <chr>, md5sum <chr>,
#> #   sha256 <chr>, filesize <int>, built <chr>, platform <chr>, rversion <chr>,
#> #   repodir <chr>, target <chr>, deps <list>, mirror <chr>, sources <list>,
#> #   remote <list>, error <list>, metadata <list>, extra <list>,
#> #   dep_types <list>, cache_status <chr>, lib_status <chr>, old_version <chr>,
#> #   new_version <chr>, fulltarget <chr>, fulltarget_tree <chr>,
#> #   download_status <chr>, download_error <list>, file_size <dbl>,
#> #   library <chr>, binary <lgl>, dependencies <I<list>>, installed <chr>,
#> #   vignettes <lgl>, packaged <lgl>, file <chr>

pdi <- new_pkg_installation_proposal(
  "pak",
  config = list(library = tempfile())
)

#> ℹ Creating library directory: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpA4pHRz/file7597bcacc76

pdi

#> <pkg_installation_proposal>
#> + refs:
#>   - pak
#> + solution policy: lazy
#> (use `$solve()` to solve dependencies)

pdi$resolve()

#> 

#> 

#> 

#>  Found  1  deps for  0/1  pkgs [⠋] Resolving standard (CRAN/BioC) packages

#>                                                                           

pdi

#> <pkg_installation_proposal>
#> + refs:
#>   - pak
#> + solution policy: lazy
#> (use `$solve()` to solve dependencies)

pdi$solve()
pdi

#> <pkg_installation_proposal>
#> + refs:
#>   - pak
#> + solution policy: lazy
#> + has solution
#> (use `$download()` to download packages)
#> (use `$show_solution()` to see the packages to install
#> (use `$get_solution()` to see the full solution results)
#> (use `$draw()` to draw the dependency tree)

pdi$download()

#> ℹ Getting 25 pkgs (7.84 MB), 18 (20.92 MB) cached

#> 

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>           ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of assertthat 0.2.1 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of base64enc 0.1-3 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of crayon 1.3.4 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of curl 4.3 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of desc 1.2.0 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of ellipsis 0.3.1 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of fansi 0.4.1 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of filelock 1.0.2 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of hms 0.5.3 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of lifecycle 0.2.0 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of lpSolve 5.6.15 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of magrittr 1.5 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of pak 0.1.2 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of pkgconfig 2.0.3 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of prettycode 1.1.0 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of prettyunits 1.1.1 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of progress 1.2.2 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of rappdirs 0.3.1 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of rematch2 2.1.2 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of rprojroot 1.3-2 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of selectr 0.4-2 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of stringr 1.4.0 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of utf8 1.1.4 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of uuid 0.1-4 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ Cached copy of xml2 1.3.2 (macos) is the latest build

#>  0.0 kB/s ⸨             ⸩   0% |  0/25 pkg | ETA ??s  | Connecting...

#>  0.0 kB/s ⸨█████████████⸩ 100% | 25/25 pkg | ETA ??s  | Connecting...

#>                                                                      

#> ✔ No downloads needed, all packages are cached

pdi

#> <pkg_installation_proposal>
#> + refs:
#>   - pak
#> + solution policy: lazy
#> + has solution
#> + has downloads
#> (use `$show_solution()` to see the packages to install
#> (use `$get_solution()` to see the full solution results)
#> (use `$draw()` to draw the dependency tree)
#> (use `$get_downloads()` to get download data)
#> (use `$get_install_plan()` to get the installation plan)
#> (use `$install()` to install the packages)